@@ -43,45 +69,50 @@
     "onCommand:PowerShell.OpenExamplesFolder",
     "onCommand:PowerShell.EnableISEMode",
     "onCommand:PowerShell.DisableISEMode",
-    "onView:PowerShellCommands"
+    "onCommand:PowerShell.ToggleISEMode",
+    "onView:PowerShellCommands",
+    "onWalkthrough:PowerShell"
   ],
   "dependencies": {
-    "@vscode/extension-telemetry": "~0.6.2",
+    "@vscode/extension-telemetry": "^0.6.2",
     "node-fetch": "~2.6.7",
-    "semver": "~7.3.7",
-    "uuid": "~8.3.2",
-    "vscode-languageclient": "~7.0.0",
+    "semver": "~7.3.8",
+    "uuid": "~9.0.0",
+    "vscode-languageclient": "~8.0.2",
     "vscode-languageserver-protocol": "~3.17.2"
   },
   "devDependencies": {
-    "@types/glob": "~7.2.0",
-    "@types/mocha": "~9.1.1",
+    "@types/glob": "~8.0.0",
+    "@types/mocha": "~10.0.1",
     "@types/mock-fs": "~4.13.1",
     "@types/node": "~14.17.4",
-    "@types/node-fetch": "~2.6.1",
+    "@types/node-fetch": "~2.6.2",
     "@types/rewire": "~2.5.28",
-    "@types/semver": "~7.3.10",
+    "@types/semver": "~7.3.13",
     "@types/sinon": "~10.0.13",
-    "@types/uuid": "~8.3.4",
-    "@types/vscode": "~1.59.0",
-    "@vscode/test-electron": "~2.1.5",
-    "esbuild": "^0.14.51",
+    "@types/uuid": "~9.0.0",
+    "@types/vscode": "~1.65.0",
+    "@typescript-eslint/eslint-plugin": "^5.46.1",
+    "@typescript-eslint/parser": "^5.46.1",
+    "@vscode/test-electron": "~2.2.1",
+    "esbuild": "^0.16.7",
+    "eslint": "^8.29.0",
+    "eslint-plugin-header": "^3.1.1",
     "glob": "^8.0.3",
-    "mocha": "~10.0.0",
+    "mocha": "~10.2.0",
     "mocha-multi-reporters": "~1.5.1",
-    "mock-fs": "~5.1.2",
+    "mock-fs": "~5.2.0",
     "rewire": "~6.0.0",
-    "sinon": "~14.0.0",
-    "tslint": "~6.1.3",
-    "typescript": "~4.7.4",
-    "vsce": "~2.10.0"
+    "sinon": "~15.0.1",
+    "typescript": "~4.9.4",
+    "vsce": "~2.15.0"
   },
   "extensionDependencies": [
     "vscode.powershell"
   ],
   "main": "./out/main.js",
   "scripts": {
-    "lint": "tslint --project tsconfig.json",
+    "lint": "eslint . --ext .ts",
     "build": "tsc --project tsconfig.json && esbuild ./src/main.ts --outdir=out --bundle --external:vscode --platform=node",
     "test": "node ./out/test/runTests.js",
     "package": "vsce package --no-gitHubIssueLinking",
