@@ -2,16 +2,20 @@
 // Licensed under the MIT License.
 
 import vscode = require("vscode");
-import { CancellationToken, DebugConfiguration, DebugConfigurationProvider,
-    ExtensionContext, WorkspaceFolder } from "vscode";
+import {
+    CancellationToken, DebugConfiguration, DebugConfigurationProvider,
+    ExtensionContext, WorkspaceFolder
+} from "vscode";
 import { NotificationType, RequestType } from "vscode-languageclient";
 import { LanguageClient } from "vscode-languageclient/node";
 import { getPlatformDetails, OperatingSystem } from "../platform";
-import { PowerShellProcess} from "../process";
+import { PowerShellProcess } from "../process";
 import { IEditorServicesSessionDetails, SessionManager, SessionStatus } from "../session";
-import Settings = require("../settings");
-import { Logger } from "../logging";
+import { getSettings } from "../settings";
+import { ILogger } from "../logging";
 import { LanguageClientConsumer } from "../languageClientConsumer";
+import path = require("path");
+import utils = require("../utils");
 
 export const StartDebuggerNotificationType =
     new NotificationType<void>("powerShell/startDebugger");
@@ -19,18 +23,53 @@ export const StartDebuggerNotificationType =
 export const StopDebuggerNotificationType =
     new NotificationType<void>("powerShell/stopDebugger");
 
+enum DebugConfig {
+    LaunchCurrentFile,
+    LaunchScript,
+    InteractiveSession,
+    AttachHostProcess,
+}
+
 export class DebugSessionFeature extends LanguageClientConsumer
     implements DebugConfigurationProvider, vscode.DebugAdapterDescriptorFactory {
 
-    private sessionCount: number = 1;
-    private tempDebugProcess: PowerShellProcess;
-    private tempSessionDetails: IEditorServicesSessionDetails;
+    private sessionCount = 1;
+    private tempDebugProcess: PowerShellProcess | undefined;
+    private tempSessionDetails: IEditorServicesSessionDetails | undefined;
+    private handlers: vscode.Disposable[] = [];
+    private configs: Record<DebugConfig, DebugConfiguration> = {
+        [DebugConfig.LaunchCurrentFile]: {
+            name: "PowerShell: Launch Current File",
+            type: "PowerShell",
+            request: "launch",
+            script: "${file}",
+            args: [],
+        },
+        [DebugConfig.LaunchScript]: {
+            name: "PowerShell: Launch Script",
+            type: "PowerShell",
+            request: "launch",
+            script: "Enter path or command to execute, for example: \"${workspaceFolder}/src/foo.ps1\" or \"Invoke-Pester\"",
+            args: [],
+        },
+        [DebugConfig.InteractiveSession]: {
+            name: "PowerShell: Interactive Session",
+            type: "PowerShell",
+            request: "launch",
+        },
+        [DebugConfig.AttachHostProcess]: {
+            name: "PowerShell: Attach to PowerShell Host Process",
+            type: "PowerShell",
+            request: "attach",
+            runspaceId: 1,
+        },
+    };
 
-    constructor(context: ExtensionContext, private sessionManager: SessionManager, private logger: Logger) {
+    constructor(context: ExtensionContext, private sessionManager: SessionManager, private logger: ILogger) {
         super();
         // Register a debug configuration provider
         context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider("PowerShell", this));
-        context.subscriptions.push(vscode.debug.registerDebugAdapterDescriptorFactory("PowerShell", this))
+        context.subscriptions.push(vscode.debug.registerDebugAdapterDescriptorFactory("PowerShell", this));
     }
 
     createDebugAdapterDescriptor(
@@ -41,41 +80,43 @@ export class DebugSessionFeature extends LanguageClientConsumer
             ? this.tempSessionDetails
             : this.sessionManager.getSessionDetails();
 
+        if (sessionDetails === undefined) {
+            void this.logger.writeAndShowError(`PowerShell session details not available for ${session.name}`);
+            return;
+        }
+
         this.logger.writeVerbose(`Connecting to pipe: ${sessionDetails.debugServicePipeName}`);
         this.logger.writeVerbose(`Debug configuration: ${JSON.stringify(session.configuration)}`);
 
         return new vscode.DebugAdapterNamedPipeServer(sessionDetails.debugServicePipeName);
     }
 
-    // tslint:disable-next-line:no-empty
     public dispose() {
+        for (const handler of this.handlers) {
+            handler.dispose();
+        }
     }
 
-    public setLanguageClient(languageClient: LanguageClient) {
-        languageClient.onNotification(
-            StartDebuggerNotificationType,
-            // TODO: Use a named debug configuration.
-            () => vscode.debug.startDebugging(undefined, {
-                request: "launch",
-                type: "PowerShell",
-                name: "PowerShell: Interactive Session"
-            }));
+    public override setLanguageClient(languageClient: LanguageClient) {
+        this.handlers = [
+            languageClient.onNotification(
+                StartDebuggerNotificationType,
+                // TODO: Use a named debug configuration.
+                () => void vscode.debug.startDebugging(undefined, {
+                    request: "launch",
+                    type: "PowerShell",
+                    name: "PowerShell: Interactive Session"
+                })),
 
-        languageClient.onNotification(
-            StopDebuggerNotificationType,
-            () => vscode.debug.stopDebugging(undefined));
+            languageClient.onNotification(
+                StopDebuggerNotificationType,
+                () => void vscode.debug.stopDebugging(undefined))
+        ];
     }
 
     public async provideDebugConfigurations(
-        folder: WorkspaceFolder | undefined,
-        token?: CancellationToken): Promise<DebugConfiguration[]> {
-
-        enum DebugConfig {
-            LaunchCurrentFile,
-            LaunchScript,
-            InteractiveSession,
-            AttachHostProcess,
-        }
+        _folder: WorkspaceFolder | undefined,
+        _token?: CancellationToken): Promise<DebugConfiguration[]> {
 
         const debugConfigPickItems = [
             {
@@ -105,213 +146,149 @@ export class DebugSessionFeature extends LanguageClientConsumer
                 debugConfigPickItems,
                 { placeHolder: "Select a PowerShell debug configuration" });
 
-        // TODO: Make these available in a dictionary and share them.
-        switch (launchSelection.id) {
-            case DebugConfig.LaunchCurrentFile:
-                return [
-                    {
-                        name: "PowerShell: Launch Current File",
-                        type: "PowerShell",
-                        request: "launch",
-                        script: "${file}",
-                        cwd: "${file}",
-                    },
-                ];
-            case DebugConfig.LaunchScript:
-                return [
-                    {
-                        name: "PowerShell: Launch Script",
-                        type: "PowerShell",
-                        request: "launch",
-                        script: "enter path or command to execute e.g.: ${workspaceFolder}/src/foo.ps1 or Invoke-Pester",
-                        cwd: "${workspaceFolder}",
-                    },
-                ];
-            case DebugConfig.InteractiveSession:
-                return [
-                    {
-                        name: "PowerShell: Interactive Session",
-                        type: "PowerShell",
-                        request: "launch",
-                        cwd: "",
-                    },
-                ];
-            case DebugConfig.AttachHostProcess:
-                return [
-                    {
-                        name: "PowerShell: Attach to PowerShell Host Process",
-                        type: "PowerShell",
-                        request: "attach",
-                        runspaceId: 1,
-                    },
-                ];
+        if (launchSelection) {
+            return [this.configs[launchSelection.id]];
         }
+
+        return [this.configs[DebugConfig.LaunchCurrentFile]];
     }
 
-    // DebugConfigurationProvider method
+    // DebugConfigurationProvider methods
     public async resolveDebugConfiguration(
         _folder: WorkspaceFolder | undefined,
         config: DebugConfiguration,
-        _token?: CancellationToken): Promise<DebugConfiguration> {
-        // Make sure there is a session running before attempting to debug/run a program
-        // TODO: Perhaps this should just wait until it's running or aborted.
-        if (this.sessionManager.getSessionStatus() !== SessionStatus.Running) {
-            const msg = "Cannot debug or run a PowerShell script until the PowerShell session has started. " +
-                "Wait for the PowerShell session to finish starting and try again.";
-            vscode.window.showWarningMessage(msg);
-            return undefined;
-        }
-
-        // Starting a debug session can be done when there is no document open e.g. attach to PS host process
-        const currentDocument = vscode.window.activeTextEditor ? vscode.window.activeTextEditor.document : undefined;
-        const debugCurrentScript = (config.script === "${file}") || !config.request;
-        const generateLaunchConfig = !config.request;
-
-        const settings = Settings.load();
-
-        // If the createTemporaryIntegratedConsole field is not specified in the launch config, set the field using
-        // the value from the corresponding setting. Otherwise, the launch config value overrides the setting.
-        config.createTemporaryIntegratedConsole =
-            config.createTemporaryIntegratedConsole ??
-            settings.debugging.createTemporaryIntegratedConsole;
-
-        if (config.request === "attach") {
-            const platformDetails = getPlatformDetails();
-            const versionDetails = this.sessionManager.getPowerShellVersionDetails();
-
-            // Cross-platform attach to process was added in 6.2.0-preview.4
-            if (versionDetails.version < "6.2.0" && platformDetails.operatingSystem !== OperatingSystem.Windows) {
-                const msg = `Attaching to a PowerShell Host Process on ${
-                    OperatingSystem[platformDetails.operatingSystem] } requires PowerShell 6.2 or higher.`;
-                return vscode.window.showErrorMessage(msg).then((_) => {
-                    return undefined;
-                });
-            }
-
-            // if nothing is set, prompt for the processId
-            if (!config.customPipeName && !config.processId) {
-                config.processId = await vscode.commands.executeCommand("PowerShell.PickPSHostProcess");
-
-                // No process selected. Cancel attach.
-                if (!config.processId) {
-                    return null;
-                }
-            }
-
-            if (!config.runspaceId && !config.runspaceName) {
-                config.runspaceId = await vscode.commands.executeCommand("PowerShell.PickRunspace", config.processId);
-
-                // No runspace selected. Cancel attach.
-                if (!config.runspaceId) {
-                    return null;
-                }
-            }
-        }
-
-        // TODO: Use a named debug configuration.
-        if (generateLaunchConfig) {
-            // No launch.json, create the default configuration for both unsaved (Untitled) and saved documents.
-            config.type = "PowerShell";
-            config.name = "PowerShell: Launch Current File";
-            config.request = "launch";
-            config.args = [];
-
-            config.script =
-                currentDocument.isUntitled
-                    ? currentDocument.uri.toString()
-                    : currentDocument.fileName;
-
-            if (config.createTemporaryIntegratedConsole) {
-                // For a folder-less workspace, vscode.workspace.rootPath will be undefined.
-                // PSES will convert that undefined to a reasonable working dir.
-                config.cwd =
-                    currentDocument.isUntitled
-                        ? vscode.workspace.rootPath
-                        : currentDocument.fileName;
-
-            } else {
-                // If the non-temp Extension Terminal is being used, default to the current working dir.
-                config.cwd = "";
-            }
-        }
-
-        if (config.request === "launch") {
-            // For debug launch of "current script" (saved or unsaved), warn before starting the debugger if either
-            // A) there is not an active document
-            // B) the unsaved document's language type is not PowerShell
-            // C) the saved document's extension is a type that PowerShell can't debug.
-            if (debugCurrentScript) {
-
-                if (currentDocument === undefined) {
-                    const msg = "To debug the \"Current File\", you must first open a " +
-                                "PowerShell script file in the editor.";
-                    vscode.window.showErrorMessage(msg);
-                    return;
-                }
-
-                if (currentDocument.isUntitled) {
-                    if (config.createTemporaryIntegratedConsole) {
-                        const msg = "Debugging Untitled files in a temporary console is currently not supported.";
-                        vscode.window.showErrorMessage(msg);
-                        return;
-                    }
-
-                    if (currentDocument.languageId === "powershell") {
-                        if (!generateLaunchConfig) {
-                            // Cover the case of existing launch.json but unsaved (Untitled) document.
-                            // In this case, vscode.workspace.rootPath will not be undefined.
-                            config.script = currentDocument.uri.toString();
-                            config.cwd = vscode.workspace.rootPath;
-                        }
-                    } else {
-                        const msg = "To debug '" + currentDocument.fileName + "', change the document's " +
-                                    "language mode to PowerShell or save the file with a PowerShell extension.";
-                        vscode.window.showErrorMessage(msg);
-                        return;
-                    }
-                } else {
-                    let isValidExtension = false;
-                    const extIndex = currentDocument.fileName.lastIndexOf(".");
-                    if (extIndex !== -1) {
-                        const ext = currentDocument.fileName.substr(extIndex + 1).toUpperCase();
-                        isValidExtension = (ext === "PS1" || ext === "PSM1");
-                    }
-
-                    if ((currentDocument.languageId !== "powershell") || !isValidExtension) {
-                        let docPath = currentDocument.fileName;
-                        const workspaceRootPath = vscode.workspace.rootPath;
-                        if (currentDocument.fileName.startsWith(workspaceRootPath)) {
-                            docPath = currentDocument.fileName.substring(vscode.workspace.rootPath.length + 1);
-                        }
-
-                        const msg = "PowerShell does not support debugging this file type: '" + docPath + "'.";
-                        vscode.window.showErrorMessage(msg);
-                        return;
-                    }
-
-                    if (config.script === "${file}") {
-                        config.script = currentDocument.fileName;
-                    }
-                }
-            }
-
-            // NOTE: There is a tight coupling to a weird setting in
-            // `package.json` for the Launch Current File configuration where
-            // the default cwd is set to ${file}.
-            if ((currentDocument !== undefined) && (config.cwd === "${file}")) {
-                config.cwd = currentDocument.fileName;
-            }
-        }
+        _token?: CancellationToken): Promise<DebugConfiguration | undefined> {
 
         // Prevent the Debug Console from opening
         config.internalConsoleOptions = "neverOpen";
 
-        // Create or show the interactive console
-        vscode.commands.executeCommand("PowerShell.ShowSessionConsole", true);
+        // NOTE: We intentionally do not touch the `cwd` setting of the config.
+
+        // If the createTemporaryIntegratedConsole field is not specified in the
+        // launch config, set the field using the value from the corresponding
+        // setting. Otherwise, the launch config value overrides the setting.
+        //
+        // Also start the temporary process and console for this configuration.
+        const settings = getSettings();
+        config.createTemporaryIntegratedConsole =
+            config.createTemporaryIntegratedConsole ??
+            settings.debugging.createTemporaryIntegratedConsole;
 
         if (config.createTemporaryIntegratedConsole) {
-            this.tempDebugProcess = this.sessionManager.createDebugSessionProcess(settings);
+            this.tempDebugProcess = await this.sessionManager.createDebugSessionProcess(settings);
             this.tempSessionDetails = await this.tempDebugProcess.start(`DebugSession-${this.sessionCount++}`);
+        }
+
+        if (!config.request) {
+            // No launch.json, create the default configuration for both unsaved
+            // (Untitled) and saved documents.
+            const LaunchCurrentFileConfig = this.configs[DebugConfig.LaunchCurrentFile];
+            config = { ...config, ...LaunchCurrentFileConfig };
+            config.current_document = true;
+        }
+
+        if (config.script === "${file}" || config.script === "${relativeFile}") {
+            if (vscode.window.activeTextEditor === undefined) {
+                void this.logger.writeAndShowError("To debug the 'Current File', you must first open a PowerShell script file in the editor.");
+                return undefined;
+            }
+            config.current_document = true;
+            // Special case using the URI for untitled documents.
+            const currentDocument = vscode.window.activeTextEditor.document;
+            if (currentDocument.isUntitled) {
+                config.untitled_document = true;
+                config.script = currentDocument.uri.toString();
+            }
+        }
+
+        return config;
+    }
+
+    public async resolveDebugConfigurationWithSubstitutedVariables(
+        _folder: WorkspaceFolder | undefined,
+        config: DebugConfiguration,
+        _token?: CancellationToken): Promise<DebugConfiguration | undefined | null> {
+
+        let resolvedConfig: DebugConfiguration | undefined | null;
+        if (config.request === "attach") {
+            resolvedConfig = await this.resolveAttachDebugConfiguration(config);
+        } else if (config.request === "launch") {
+            resolvedConfig = await this.resolveLaunchDebugConfiguration(config);
+        } else {
+            void this.logger.writeAndShowError(`PowerShell debug configuration's request type was invalid: '${config.request}'.`);
+            return null;
+        }
+
+        if (resolvedConfig) {
+            // Start the PowerShell session if needed.
+            if (this.sessionManager.getSessionStatus() !== SessionStatus.Running) {
+                await this.sessionManager.start();
+            }
+            // Create or show the debug terminal (either temporary or session).
+            this.sessionManager.showDebugTerminal(true);
+        }
+
+        return resolvedConfig;
+    }
+
+    private async resolveLaunchDebugConfiguration(config: DebugConfiguration): Promise<DebugConfiguration | undefined> {
+        // Check the languageId and file extension only for current documents
+        // (which includes untitled documents). This prevents accidentally
+        // running the debugger for an open non-PowerShell file.
+        if (config.current_document) {
+            const currentDocument = vscode.window.activeTextEditor?.document;
+            if (currentDocument?.languageId !== "powershell") {
+                void this.logger.writeAndShowError(`PowerShell does not support debugging this language mode: '${currentDocument?.languageId}'.`);
+                return undefined;
+            }
+
+            if (await utils.checkIfFileExists(config.script)) {
+                const ext = path.extname(config.script).toLowerCase();
+                if (!(ext === ".ps1" || ext === ".psm1")) {
+                    void this.logger.writeAndShowError(`PowerShell does not support debugging this file type: '${path.basename(config.script)}'.`);
+                    return undefined;
+                }
+            }
+        }
+
+        // Check the temporary console setting for untitled documents only.
+        if (config.untitled_document && config.createTemporaryIntegratedConsole) {
+            void this.logger.writeAndShowError("PowerShell does not support debugging untitled files in a temporary console.");
+            return undefined;
+        }
+
+        return config;
+    }
+
+    private async resolveAttachDebugConfiguration(config: DebugConfiguration): Promise<DebugConfiguration | undefined | null> {
+        const platformDetails = getPlatformDetails();
+        const versionDetails = this.sessionManager.getPowerShellVersionDetails();
+        if (versionDetails === undefined) {
+            void this.logger.writeAndShowError(`PowerShell session version details were not found for '${config.name}'.`);
+            return null;
+        }
+
+        // Cross-platform attach to process was added in 6.2.0-preview.4.
+        if (versionDetails.version < "7.0.0" && platformDetails.operatingSystem !== OperatingSystem.Windows) {
+            void this.logger.writeAndShowError(`Attaching to a PowerShell Host Process on ${OperatingSystem[platformDetails.operatingSystem]} requires PowerShell 7.0 or higher.`);
+            return undefined;
+        }
+
+        // If nothing is set, prompt for the processId.
+        if (!config.customPipeName && !config.processId) {
+            config.processId = await vscode.commands.executeCommand("PowerShell.PickPSHostProcess");
+            // No process selected. Cancel attach.
+            if (!config.processId) {
+                return null;
+            }
+        }
+
+        if (!config.runspaceId && !config.runspaceName) {
+            config.runspaceId = await vscode.commands.executeCommand("PowerShell.PickRunspace", config.processId);
+            // No runspace selected. Cancel attach.
+            if (!config.runspaceId) {
+                return null;
+            }
         }
 
         return config;
@@ -326,17 +303,16 @@ export class SpecifyScriptArgsFeature implements vscode.Disposable {
     constructor(context: vscode.ExtensionContext) {
         this.context = context;
 
-        this.command =
-            vscode.commands.registerCommand("PowerShell.SpecifyScriptArgs", () => {
-                return this.specifyScriptArguments();
-            });
+        this.command = vscode.commands.registerCommand("PowerShell.SpecifyScriptArgs", () => {
+            return this.specifyScriptArguments();
+        });
     }
 
     public dispose() {
         this.command.dispose();
     }
 
-    private async specifyScriptArguments(): Promise<string> {
+    private async specifyScriptArguments(): Promise<string | undefined> {
         const powerShellDbgScriptArgsKey = "powerShellDebugScriptArgs";
 
         const options: vscode.InputBoxOptions = {
@@ -353,7 +329,7 @@ export class SpecifyScriptArgsFeature implements vscode.Disposable {
         // When user cancel's the input box (by pressing Esc), the text value is undefined.
         // Let's not blow away the previous setting.
         if (text !== undefined) {
-            this.context.workspaceState.update(powerShellDbgScriptArgsKey, text);
+            await this.context.workspaceState.update(powerShellDbgScriptArgsKey, text);
         }
         return text;
     }
@@ -361,6 +337,10 @@ export class SpecifyScriptArgsFeature implements vscode.Disposable {
 
 interface IProcessItem extends vscode.QuickPickItem {
     pid: string;    // payload for the QuickPick UI
+}
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+interface IGetPSHostProcessesArguments {
 }
 
 interface IPSHostProcessInfo {
@@ -371,32 +351,28 @@ interface IPSHostProcessInfo {
 }
 
 export const GetPSHostProcessesRequestType =
-    new RequestType<any, IGetPSHostProcessesResponseBody, string>("powerShell/getPSHostProcesses");
-
-interface IGetPSHostProcessesResponseBody {
-    hostProcesses: IPSHostProcessInfo[];
-}
+    new RequestType<IGetPSHostProcessesArguments, IPSHostProcessInfo[], string>("powerShell/getPSHostProcesses");
 
 export class PickPSHostProcessFeature extends LanguageClientConsumer {
 
     private command: vscode.Disposable;
-    private waitingForClientToken: vscode.CancellationTokenSource;
-    private getLanguageClientResolve: (value?: LanguageClient | Promise<LanguageClient>) => void;
+    private waitingForClientToken?: vscode.CancellationTokenSource;
+    private getLanguageClientResolve?: (value: LanguageClient) => void;
 
-    constructor() {
+    constructor(private logger: ILogger) {
         super();
 
         this.command =
             vscode.commands.registerCommand("PowerShell.PickPSHostProcess", () => {
                 return this.getLanguageClient()
-                           .then((_) => this.pickPSHostProcess(), (_) => undefined);
+                    .then((_) => this.pickPSHostProcess(), (_) => undefined);
             });
     }
 
-    public setLanguageClient(languageClient: LanguageClient) {
+    public override setLanguageClient(languageClient: LanguageClient) {
         this.languageClient = languageClient;
 
-        if (this.waitingForClientToken) {
+        if (this.waitingForClientToken && this.getLanguageClientResolve) {
             this.getLanguageClientResolve(this.languageClient);
             this.clearWaitingToken();
         }
@@ -407,7 +383,7 @@ export class PickPSHostProcessFeature extends LanguageClientConsumer {
     }
 
     private getLanguageClient(): Promise<LanguageClient> {
-        if (this.languageClient) {
+        if (this.languageClient !== undefined) {
             return Promise.resolve(this.languageClient);
         } else {
             // If PowerShell isn't finished loading yet, show a loading message
@@ -422,13 +398,13 @@ export class PickPSHostProcessFeature extends LanguageClientConsumer {
                         .showQuickPick(
                             ["Cancel"],
                             { placeHolder: "Attach to PowerShell host process: Please wait, starting PowerShell..." },
-                            this.waitingForClientToken.token)
+                            this.waitingForClientToken?.token)
                         .then((response) => {
                             if (response === "Cancel") {
                                 this.clearWaitingToken();
                                 reject();
                             }
-                        });
+                        }, undefined);
 
                     // Cancel the loading prompt after 60 seconds
                     setTimeout(() => {
@@ -436,8 +412,7 @@ export class PickPSHostProcessFeature extends LanguageClientConsumer {
                             this.clearWaitingToken();
                             reject();
 
-                            vscode.window.showErrorMessage(
-                                "Attach to PowerShell host process: PowerShell session took too long to start.");
+                            void this.logger.writeAndShowError("Attach to PowerShell host process: PowerShell session took too long to start.");
                         }
                     }, 60000);
                 },
@@ -445,50 +420,54 @@ export class PickPSHostProcessFeature extends LanguageClientConsumer {
         }
     }
 
-    private async pickPSHostProcess(): Promise<string> {
-        const hostProcesses = await this.languageClient.sendRequest(GetPSHostProcessesRequestType, {});
+    private async pickPSHostProcess(): Promise<string | undefined> {
         // Start with the current PowerShell process in the list.
         const items: IProcessItem[] = [{
             label: "Current",
             description: "The current PowerShell Extension process.",
             pid: "current",
         }];
-        for (const p in hostProcesses) {
-            if (hostProcesses.hasOwnProperty(p)) {
-                let windowTitle = "";
-                if (hostProcesses[p].mainWindowTitle) {
-                    windowTitle = `, Title: ${hostProcesses[p].mainWindowTitle}`;
-                }
 
-                items.push({
-                    label: hostProcesses[p].processName,
-                    description: `PID: ${hostProcesses[p].processId.toString()}${windowTitle}`,
-                    pid: hostProcesses[p].processId,
-                });
+        const response = await this.languageClient?.sendRequest(GetPSHostProcessesRequestType, {});
+        for (const process of response ?? []) {
+            let windowTitle = "";
+            if (process.mainWindowTitle) {
+                windowTitle = `, Title: ${process.mainWindowTitle}`;
             }
+
+            items.push({
+                label: process.processName,
+                description: `PID: ${process.processId.toString()}${windowTitle}`,
+                pid: process.processId,
+            });
         }
+
         if (items.length === 0) {
             return Promise.reject("There are no PowerShell host processes to attach to.");
         }
+
         const options: vscode.QuickPickOptions = {
             placeHolder: "Select a PowerShell host process to attach to",
             matchOnDescription: true,
             matchOnDetail: true,
         };
         const item = await vscode.window.showQuickPick(items, options);
+
         return item ? `${item.pid}` : undefined;
     }
 
     private clearWaitingToken() {
-        if (this.waitingForClientToken) {
-            this.waitingForClientToken.dispose();
-            this.waitingForClientToken = undefined;
-        }
+        this.waitingForClientToken?.dispose();
+        this.waitingForClientToken = undefined;
     }
 }
 
 interface IRunspaceItem extends vscode.QuickPickItem {
     id: string;    // payload for the QuickPick UI
+}
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+interface IGetRunspaceRequestArguments {
 }
 
 interface IRunspace {
@@ -498,27 +477,27 @@ interface IRunspace {
 }
 
 export const GetRunspaceRequestType =
-    new RequestType<any, IRunspace[], string>("powerShell/getRunspace");
+    new RequestType<IGetRunspaceRequestArguments, IRunspace[], string>("powerShell/getRunspace");
 
 export class PickRunspaceFeature extends LanguageClientConsumer {
 
     private command: vscode.Disposable;
-    private waitingForClientToken: vscode.CancellationTokenSource;
-    private getLanguageClientResolve: (value?: LanguageClient | Promise<LanguageClient>) => void;
+    private waitingForClientToken?: vscode.CancellationTokenSource;
+    private getLanguageClientResolve?: (value: LanguageClient) => void;
 
-    constructor() {
+    constructor(private logger: ILogger) {
         super();
         this.command =
             vscode.commands.registerCommand("PowerShell.PickRunspace", (processId) => {
                 return this.getLanguageClient()
-                           .then((_) => this.pickRunspace(processId), (_) => undefined);
+                    .then((_) => this.pickRunspace(processId), (_) => undefined);
             }, this);
     }
 
-    public setLanguageClient(languageClient: LanguageClient) {
+    public override setLanguageClient(languageClient: LanguageClient) {
         this.languageClient = languageClient;
 
-        if (this.waitingForClientToken) {
+        if (this.waitingForClientToken && this.getLanguageClientResolve) {
             this.getLanguageClientResolve(this.languageClient);
             this.clearWaitingToken();
         }
@@ -544,13 +523,13 @@ export class PickRunspaceFeature extends LanguageClientConsumer {
                         .showQuickPick(
                             ["Cancel"],
                             { placeHolder: "Attach to PowerShell host process: Please wait, starting PowerShell..." },
-                            this.waitingForClientToken.token)
+                            this.waitingForClientToken?.token)
                         .then((response) => {
                             if (response === "Cancel") {
                                 this.clearWaitingToken();
                                 reject();
                             }
-                        });
+                        }, undefined);
 
                     // Cancel the loading prompt after 60 seconds
                     setTimeout(() => {
@@ -558,8 +537,7 @@ export class PickRunspaceFeature extends LanguageClientConsumer {
                             this.clearWaitingToken();
                             reject();
 
-                            vscode.window.showErrorMessage(
-                                "Attach to PowerShell host process: PowerShell session took too long to start.");
+                            void this.logger.writeAndShowError("Attach to PowerShell host process: PowerShell session took too long to start.");
                         }
                     }, 60000);
                 },
@@ -567,10 +545,10 @@ export class PickRunspaceFeature extends LanguageClientConsumer {
         }
     }
 
-    private async pickRunspace(processId: string): Promise<string> {
-        const response = await this.languageClient.sendRequest(GetRunspaceRequestType, { processId });
+    private async pickRunspace(processId: string): Promise<string | undefined> {
+        const response = await this.languageClient?.sendRequest(GetRunspaceRequestType, { processId });
         const items: IRunspaceItem[] = [];
-        for (const runspace of response) {
+        for (const runspace of response ?? []) {
             // Skip default runspace
             if ((runspace.id === 1 || runspace.name === "PSAttachRunspace")
                 && processId === "current") {
@@ -583,19 +561,19 @@ export class PickRunspaceFeature extends LanguageClientConsumer {
                 id: runspace.id.toString(),
             });
         }
+
         const options: vscode.QuickPickOptions = {
             placeHolder: "Select PowerShell runspace to debug",
             matchOnDescription: true,
             matchOnDetail: true,
         };
         const item = await vscode.window.showQuickPick(items, options);
+
         return item ? `${item.id}` : undefined;
     }
 
     private clearWaitingToken() {
-        if (this.waitingForClientToken) {
-            this.waitingForClientToken.dispose();
-            this.waitingForClientToken = undefined;
-        }
+        this.waitingForClientToken?.dispose();
+        this.waitingForClientToken = undefined;
     }
 }
