@@ -3,8 +3,9 @@
 
 import * as path from "path";
 import vscode = require("vscode");
+import { ILogger } from "../logging";
 import { SessionManager } from "../session";
-import Settings = require("../settings");
+import { getSettings, chosenWorkspace, validateCwdSetting } from "../settings";
 import utils = require("../utils");
 
 enum LaunchType {
@@ -13,81 +14,85 @@ enum LaunchType {
 }
 
 export class PesterTestsFeature implements vscode.Disposable {
-
-    private command: vscode.Disposable;
+    private commands: vscode.Disposable[];
     private invokePesterStubScriptPath: string;
 
-    constructor(private sessionManager: SessionManager) {
+    constructor(private sessionManager: SessionManager, private logger: ILogger) {
         this.invokePesterStubScriptPath = path.resolve(__dirname, "../modules/PowerShellEditorServices/InvokePesterStub.ps1");
+        this.commands = [
+            // File context-menu command - Run Pester Tests
+            vscode.commands.registerCommand(
+                "PowerShell.RunPesterTestsFromFile",
+                (fileUri?) => {
+                    return this.launchAllTestsInActiveEditor(LaunchType.Run, fileUri);
+                }),
 
-        // File context-menu command - Run Pester Tests
-        this.command = vscode.commands.registerCommand(
-            "PowerShell.RunPesterTestsFromFile",
-            (fileUri) => {
-                return this.launchAllTestsInActiveEditor(LaunchType.Run, fileUri);
-            });
-        // File context-menu command - Debug Pester Tests
-        this.command = vscode.commands.registerCommand(
-            "PowerShell.DebugPesterTestsFromFile",
-            (fileUri) => {
-                return this.launchAllTestsInActiveEditor(LaunchType.Debug, fileUri);
-            });
-        // This command is provided for usage by PowerShellEditorServices (PSES) only
-        this.command = vscode.commands.registerCommand(
-            "PowerShell.RunPesterTests",
-            (uriString, runInDebugger, describeBlockName?, describeBlockLineNumber?, outputPath?) => {
-                return this.launchTests(uriString, runInDebugger, describeBlockName, describeBlockLineNumber, outputPath);
-            });
+            // File context-menu command - Debug Pester Tests
+            vscode.commands.registerCommand(
+                "PowerShell.DebugPesterTestsFromFile",
+                (fileUri?) => {
+                    return this.launchAllTestsInActiveEditor(LaunchType.Debug, fileUri);
+                }),
+
+            // This command is provided for usage by PowerShellEditorServices (PSES) only
+            vscode.commands.registerCommand(
+                "PowerShell.RunPesterTests",
+                (uriString, runInDebugger, describeBlockName?, describeBlockLineNumber?, outputPath?) => {
+                    return this.launchTests(vscode.Uri.parse(uriString), runInDebugger, describeBlockName, describeBlockLineNumber, outputPath);
+                })
+        ];
     }
 
     public dispose() {
-        this.command.dispose();
+        for (const command of this.commands) {
+            command.dispose();
+        }
     }
 
     private async launchAllTestsInActiveEditor(
         launchType: LaunchType,
-        fileUri: vscode.Uri): Promise<boolean> {
+        fileUri?: vscode.Uri): Promise<boolean> {
 
-        const uriString = (fileUri || vscode.window.activeTextEditor.document.uri).toString();
-        const launchConfig = this.createLaunchConfig(uriString, launchType);
+        if (fileUri === undefined) {
+            fileUri = vscode.window.activeTextEditor?.document.uri;
+        }
+
+        if (fileUri === undefined) {
+            return false;
+        }
+
+        const launchConfig = this.createLaunchConfig(fileUri, launchType);
         return this.launch(launchConfig);
     }
 
     private async launchTests(
-        uriString: string,
+        fileUri: vscode.Uri,
         runInDebugger: boolean,
         describeBlockName?: string,
         describeBlockLineNumber?: number,
         outputPath?: string): Promise<boolean> {
 
         const launchType = runInDebugger ? LaunchType.Debug : LaunchType.Run;
-        const launchConfig = this.createLaunchConfig(uriString, launchType, describeBlockName, describeBlockLineNumber, outputPath);
+        const launchConfig = this.createLaunchConfig(fileUri, launchType, describeBlockName, describeBlockLineNumber, outputPath);
         return this.launch(launchConfig);
     }
 
     private createLaunchConfig(
-        uriString: string,
+        fileUri: vscode.Uri,
         launchType: LaunchType,
         testName?: string,
         lineNum?: number,
-        outputPath?: string) {
+        outputPath?: string): vscode.DebugConfiguration {
 
-        const uri = vscode.Uri.parse(uriString);
-        const currentDocument = vscode.window.activeTextEditor.document;
-        const settings = Settings.load();
-
-        // Since we pass the script path to PSES in single quotes to avoid issues with PowerShell
-        // special chars like & $ @ () [], we do have to double up the interior single quotes.
-        const scriptPath = uri.fsPath.replace(/'/g, "''");
-
+        const settings = getSettings();
         const launchConfig = {
             request: "launch",
             type: "PowerShell",
-            name: "PowerShell Launch Pester Tests",
+            name: "PowerShell: Launch Pester Tests",
             script: this.invokePesterStubScriptPath,
             args: [
                 "-ScriptPath",
-                `'${scriptPath}'`,
+                `'${utils.escapeSingleQuotes(fileUri.fsPath)}'`,
             ],
             internalConsoleOptions: "neverOpen",
             noDebug: (launchType === LaunchType.Run),
@@ -97,12 +102,7 @@ export class PesterTestsFeature implements vscode.Disposable {
         if (lineNum) {
             launchConfig.args.push("-LineNumber", `${lineNum}`);
         } else if (testName) {
-            // Escape single quotes inside double quotes by doubling them up
-            if (testName.includes("'")) {
-                testName = testName.replace(/'/g, "''");
-            }
-
-            launchConfig.args.push("-TestName", `'${testName}'`);
+            launchConfig.args.push("-TestName", `'${utils.escapeSingleQuotes(testName)}'`);
         } else {
             launchConfig.args.push("-All");
         }
@@ -125,16 +125,15 @@ export class PesterTestsFeature implements vscode.Disposable {
         return launchConfig;
     }
 
-    private async launch(launchConfig): Promise<boolean> {
+    private async launch(launchConfig: vscode.DebugConfiguration): Promise<boolean> {
         // Create or show the interactive console
         // TODO: #367 Check if "newSession" mode is configured
-        await vscode.commands.executeCommand("PowerShell.ShowSessionConsole", true);
+        this.sessionManager.showDebugTerminal(true);
 
-        // TODO: Update to handle multiple root workspaces.
-        //
         // Ensure the necessary script exists (for testing). The debugger will
         // start regardless, but we also pass its success along.
-        return utils.fileExists(this.invokePesterStubScriptPath)
-            && vscode.debug.startDebugging(vscode.workspace.workspaceFolders?.[0], launchConfig);
+        await validateCwdSetting(this.logger);
+        return await utils.checkIfFileExists(this.invokePesterStubScriptPath)
+            && vscode.debug.startDebugging(chosenWorkspace, launchConfig);
     }
 }
