@@ -3,47 +3,34 @@
 
 import * as assert from "assert";
 import * as vscode from "vscode";
-import Settings = require("../../src/settings");
+import * as settings from "../../src/settings";
 
 describe("Settings module", function () {
     it("Loads without error", function () {
-        assert.doesNotThrow(Settings.load);
+        assert.doesNotThrow(settings.getSettings);
     });
 
-    it("Updates correctly with 'then' syntax", async function () {
-        Settings.change("helpCompletion", "BlockComment", false).then(() =>
-            assert.strictEqual(Settings.load().helpCompletion, "BlockComment"));
+    it("Loads the correct defaults", function () {
+        const testSettings = new settings.Settings();
+        testSettings.enableProfileLoading = false;
+        testSettings.powerShellAdditionalExePaths = { "Some PowerShell": "somePath" };
+        const actualSettings = settings.getSettings();
+        assert.deepStrictEqual(actualSettings, testSettings);
     });
 
-    it("Updates correctly with 'async/await' syntax", async function () {
-        await Settings.change("helpCompletion", "LineComment", false);
-        assert.strictEqual(Settings.load().helpCompletion, "LineComment");
-    });
 
-    describe("User-only settings", async function () {
-        const psExeDetails = {
-            "My PowerShell": "dummyPath",
-        };
-
-        it("Throws when updating at workspace-level", async function () {
-            assert.rejects(async () => await Settings.change("powerShellAdditionalExePaths", psExeDetails, false /* workspace-level */));
-        });
-
-        it("Doesn't throw when updating at user-level", async function () {
-            await Settings.change("powerShellAdditionalExePaths", psExeDetails, true /* user-level */);
-            const result = Settings.load().powerShellAdditionalExePaths["My PowerShell"];
-            assert.notStrictEqual(result, undefined);
-            assert.strictEqual(result, psExeDetails["My PowerShell"]);
-        });
+    it("Updates correctly", async function () {
+        await settings.changeSetting("helpCompletion", settings.CommentType.LineComment, false, undefined);
+        assert.strictEqual(settings.getSettings().helpCompletion, settings.CommentType.LineComment);
     });
 
     it("Gets the effective configuration target", async function () {
-        await Settings.change("helpCompletion", "LineComment", false);
-        let target = await Settings.getEffectiveConfigurationTarget("helpCompletion");
+        await settings.changeSetting("helpCompletion", settings.CommentType.LineComment, false, undefined);
+        let target = settings.getEffectiveConfigurationTarget("helpCompletion");
         assert.strictEqual(target, vscode.ConfigurationTarget.Workspace);
 
-        await Settings.change("helpCompletion", undefined, false);
-        target = await Settings.getEffectiveConfigurationTarget("helpCompletion");
-        assert.strictEqual(target, null);
+        await settings.changeSetting("helpCompletion", undefined, false, undefined);
+        target = settings.getEffectiveConfigurationTarget("helpCompletion");
+        assert.strictEqual(target, undefined);
     });
 });
