@@ -3,10 +3,10 @@
 
 import vscode = require("vscode");
 
-const confirmItemLabel: string = "$(checklist) Confirm";
-const checkedPrefix: string = "[ $(check) ]";
-const uncheckedPrefix: string = "[     ]";
-const defaultPlaceHolder: string = "Select 'Confirm' to confirm or press 'Esc' key to cancel";
+const confirmItemLabel = "$(checklist) Confirm";
+const checkedPrefix = "[ $(check) ]";
+const uncheckedPrefix = "[     ]";
+const defaultPlaceHolder = "Select 'Confirm' to confirm or press 'Esc' key to cancel";
 
 export interface ICheckboxQuickPickItem {
     label: string;
@@ -18,69 +18,58 @@ export interface ICheckboxQuickPickOptions {
     confirmPlaceHolder: string;
 }
 
-const defaultOptions: ICheckboxQuickPickOptions = { confirmPlaceHolder: defaultPlaceHolder};
+const defaultOptions: ICheckboxQuickPickOptions = { confirmPlaceHolder: defaultPlaceHolder };
 
-export function showCheckboxQuickPick(
+export async function showCheckboxQuickPick(
     items: ICheckboxQuickPickItem[],
-    options: ICheckboxQuickPickOptions = defaultOptions): Thenable<ICheckboxQuickPickItem[]> {
+    options: ICheckboxQuickPickOptions = defaultOptions): Promise<ICheckboxQuickPickItem[] | undefined> {
 
-    return showInner(items, options).then(
-        (selectedItem) => {
-            // We're mutating the original item list so just return it for now.
-            // If 'selectedItem' is undefined it means the user cancelled the
-            // inner showQuickPick UI so pass the undefined along.
-            return selectedItem !== undefined ? items : undefined;
-        });
+    const selectedItem = await showInner(items, options);
+    return selectedItem !== undefined ? items : undefined;
 }
 
 function getQuickPickItems(items: ICheckboxQuickPickItem[]): vscode.QuickPickItem[] {
-
     const quickPickItems: vscode.QuickPickItem[] = [];
     quickPickItems.push({ label: confirmItemLabel, description: "" });
 
-    items.forEach((item) =>
+    for (const item of items) {
         quickPickItems.push({
-                label: convertToCheckBox(item),
-                description: item.description,
-        }));
+            label: convertToCheckBox(item),
+            description: item.description,
+        });
+    }
 
     return quickPickItems;
 }
 
-function showInner(
+async function showInner(
     items: ICheckboxQuickPickItem[],
-    options: ICheckboxQuickPickOptions): Thenable<vscode.QuickPickItem> {
+    options: ICheckboxQuickPickOptions): Promise<vscode.QuickPickItem | undefined> {
 
-    const quickPickThenable: Thenable<vscode.QuickPickItem> =
-        vscode.window.showQuickPick(
-            getQuickPickItems(items),
-            {
-                ignoreFocusOut: true,
-                matchOnDescription: true,
-                placeHolder: options.confirmPlaceHolder,
-            });
-
-    return quickPickThenable.then(
-        (selection) => {
-            if (!selection) {
-                return Promise.resolve<vscode.QuickPickItem>(undefined);
-            }
-
-            if (selection.label === confirmItemLabel) {
-                return selection;
-            }
-
-            const index: number = getItemIndex(items, selection.label);
-
-            if (index >= 0) {
-                toggleSelection(items[index]);
-            } else {
-                // tslint:disable-next-line:no-console
-                console.log(`Couldn't find CheckboxQuickPickItem for label '${selection.label}'`);
-            }
-
-            return showInner(items, options);
+    const selection = await vscode.window.showQuickPick(
+        getQuickPickItems(items),
+        {
+            ignoreFocusOut: true,
+            matchOnDescription: true,
+            placeHolder: options.confirmPlaceHolder,
         });
+
+    if (selection === undefined) {
+        return undefined;
+    }
+
+    if (selection.label === confirmItemLabel) {
+        return selection;
+    }
+
+    const index: number = getItemIndex(items, selection.label);
+    if (index >= 0) {
+        toggleSelection(items[index]);
+    } else {
+        console.log(`Couldn't find CheckboxQuickPickItem for label '${selection.label}'`);
+    }
+
+    return showInner(items, options);
 }
 
 function getItemIndex(items: ICheckboxQuickPickItem[], itemLabel: string): number {
