@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 import * as vscode from "vscode";
-import * as Settings from "../settings";
+import { getSettings } from "../settings";
 
 interface ISetting {
     path: string;
@@ -25,25 +25,35 @@ export class ISECompatibilityFeature implements vscode.Disposable {
         { path: "editor", name: "wordSeparators", value: "`~!@#%^&*()-=+[{]}\\|;:'\",.<>/?" },
         { path: "powershell.buttons", name: "showPanelMovementButtons", value: true }
     ];
-    private iseCommandRegistration: vscode.Disposable;
-    private defaultCommandRegistration: vscode.Disposable;
+
+    private _commandRegistrations: vscode.Disposable[] = [];
+    private _iseModeEnabled: boolean;
+    private _originalSettings: Record<string, boolean | string | undefined> = {};
 
     constructor() {
-        this.iseCommandRegistration = vscode.commands.registerCommand(
-            "PowerShell.EnableISEMode", this.EnableISEMode);
-        this.defaultCommandRegistration = vscode.commands.registerCommand(
-            "PowerShell.DisableISEMode", this.DisableISEMode);
+        // TODO: This test isn't great.
+        const testSetting = ISECompatibilityFeature.settings[ISECompatibilityFeature.settings.length - 1];
+        this._iseModeEnabled = vscode.workspace.getConfiguration(testSetting.path).get(testSetting.name) === testSetting.value;
+        this._commandRegistrations = [
+            vscode.commands.registerCommand("PowerShell.EnableISEMode", async () => { await this.EnableISEMode(); }),
+            vscode.commands.registerCommand("PowerShell.DisableISEMode", async () => { await this.DisableISEMode(); }),
+            vscode.commands.registerCommand("PowerShell.ToggleISEMode", async () => { await this.ToggleISEMode(); })
+        ];
     }
 
     public dispose() {
-        this.iseCommandRegistration.dispose();
-        this.defaultCommandRegistration.dispose();
+        for (const command of this._commandRegistrations) {
+            command.dispose();
+        }
     }
 
     private async EnableISEMode() {
+        this._iseModeEnabled = true;
         for (const iseSetting of ISECompatibilityFeature.settings) {
             try {
-                await vscode.workspace.getConfiguration(iseSetting.path).update(iseSetting.name, iseSetting.value, true);
+                const config = vscode.workspace.getConfiguration(iseSetting.path);
+                this._originalSettings[iseSetting.path + iseSetting.name] = config.get(iseSetting.name);
+                await config.update(iseSetting.name, iseSetting.value, true);
             } catch {
                 // The `update` call can fail if the setting doesn't exist. This
                 // happens when the extension runs in Azure Data Studio, which
@@ -56,18 +66,28 @@ export class ISECompatibilityFeature implements vscode.Disposable {
         // Show the PowerShell view container which has the Command Explorer view
         await vscode.commands.executeCommand("workbench.view.extension.PowerShell");
 
-        if (!Settings.load().sideBar.CommandExplorerVisibility) {
+        if (!getSettings().sideBar.CommandExplorerVisibility) {
             // Hide the explorer if the setting says so.
             await vscode.commands.executeCommand("workbench.action.toggleSidebarVisibility");
         }
     }
 
     private async DisableISEMode() {
+        this._iseModeEnabled = false;
         for (const iseSetting of ISECompatibilityFeature.settings) {
-            const currently = vscode.workspace.getConfiguration(iseSetting.path).get<string | boolean>(iseSetting.name);
+            const config = vscode.workspace.getConfiguration(iseSetting.path);
+            const currently = config.get(iseSetting.name);
             if (currently === iseSetting.value) {
-                await vscode.workspace.getConfiguration(iseSetting.path).update(iseSetting.name, undefined, true);
+                await config.update(iseSetting.name, this._originalSettings[iseSetting.path + iseSetting.name], true);
             }
+        }
+    }
+
+    private async ToggleISEMode() {
+        if (this._iseModeEnabled) {
+            await this.DisableISEMode();
+        } else {
+            await this.EnableISEMode();
         }
     }
 }
