@@ -9,14 +9,16 @@ export class OpenInISEFeature implements vscode.Disposable {
 
     constructor() {
         this.command = vscode.commands.registerCommand("PowerShell.OpenInISE", () => {
-
             const editor = vscode.window.activeTextEditor;
+            if (editor === undefined) {
+                return;
+            }
+
             const document = editor.document;
             const uri = document.uri;
-
             let ISEPath = process.env.windir;
 
-            if (process.env.hasOwnProperty("PROCESSOR_ARCHITEW6432")) {
+            if (process.env.PROCESSOR_ARCHITEW6432 !== undefined) {
                 ISEPath += "\\Sysnative";
             } else {
                 ISEPath += "\\System32";
@@ -24,7 +26,7 @@ export class OpenInISEFeature implements vscode.Disposable {
 
             ISEPath += "\\WindowsPowerShell\\v1.0\\powershell_ise.exe";
 
-            ChildProcess.exec(ISEPath + ` -File "${uri.fsPath}"`).unref();
+            ChildProcess.exec(`${ISEPath} -File "${uri.fsPath}"`).unref();
         });
     }
 
