@@ -22,7 +22,7 @@ export const DidSaveTextDocumentNotificationType =
         "textDocument/didSave");
 
 export class RemoteFilesFeature extends LanguageClientConsumer {
-
+    private command: vscode.Disposable;
     private tempSessionPathPrefix: string;
 
     constructor() {
@@ -36,9 +36,9 @@ export class RemoteFilesFeature extends LanguageClientConsumer {
         // At startup, close any lingering temporary remote files
         this.closeRemoteFiles();
 
-        vscode.workspace.onDidSaveTextDocument((doc) => {
+        this.command = vscode.workspace.onDidSaveTextDocument(async (doc) => {
             if (this.isDocumentRemote(doc) && this.languageClient) {
-                this.languageClient.sendNotification(
+                await this.languageClient.sendNotification(
                     DidSaveTextDocumentNotificationType,
                     {
                         textDocument: TextDocumentIdentifier.create(doc.uri.toString()),
@@ -48,6 +48,7 @@ export class RemoteFilesFeature extends LanguageClientConsumer {
     }
 
     public dispose() {
+        this.command.dispose();
         // Close any leftover remote files before exiting
         this.closeRemoteFiles();
     }
@@ -60,17 +61,17 @@ export class RemoteFilesFeature extends LanguageClientConsumer {
         const remoteDocuments =
             vscode.workspace.textDocuments.filter((doc) => this.isDocumentRemote(doc));
 
-        function innerCloseFiles(): Thenable<{}> {
-            if (remoteDocuments.length > 0) {
-                const doc = remoteDocuments.pop();
-
-                return vscode.window
-                    .showTextDocument(doc)
-                    .then((editor) => vscode.commands.executeCommand("workbench.action.closeActiveEditor"))
-                    .then((_) => innerCloseFiles());
+        async function innerCloseFiles(): Promise<void> {
+            const doc = remoteDocuments.pop();
+            if (doc === undefined) {
+                return;
             }
+
+            await vscode.window.showTextDocument(doc);
+            await vscode.commands.executeCommand("workbench.action.closeActiveEditor");
+            return await innerCloseFiles();
         }
 
-        innerCloseFiles();
+        void innerCloseFiles();
     }
 }
