@@ -1,21 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import fs = require("fs");
 import cp = require("child_process");
-import * as semver from "semver";
 import path = require("path");
 import vscode = require("vscode");
-import { Logger } from "./logging";
+import { ILogger } from "./logging";
 import Settings = require("./settings");
 import utils = require("./utils");
-import { IEditorServicesSessionDetails, SessionManager } from "./session";
+import { IEditorServicesSessionDetails } from "./session";
 
 export class PowerShellProcess {
-    public static escapeSingleQuotes(psPath: string): string {
-        return psPath.replace(new RegExp("'", "g"), "''");
-    }
-
     // This is used to warn the user that the extension is taking longer than expected to startup.
     // After the 15th try we've hit 30 seconds and should warn.
     private static warnUserThreshold = 15;
@@ -23,23 +17,23 @@ export class PowerShellProcess {
     public onExited: vscode.Event<void>;
     private onExitedEmitter = new vscode.EventEmitter<void>();
 
-    private consoleTerminal: vscode.Terminal = undefined;
-    private consoleCloseSubscription: vscode.Disposable;
+    private consoleTerminal?: vscode.Terminal;
+    private consoleCloseSubscription?: vscode.Disposable;
 
     constructor(
         public exePath: string,
         private bundledModulesPath: string,
         private title: string,
-        private log: Logger,
+        private logger: ILogger,
         private startPsesArgs: string,
         private sessionFilePath: vscode.Uri,
-        private sessionSettings: Settings.ISettings) {
+        private sessionSettings: Settings.Settings) {
 
         this.onExited = this.onExitedEmitter.event;
     }
 
     public async start(logFileName: string): Promise<IEditorServicesSessionDetails> {
-        const editorServicesLogPath = this.log.getLogFilePath(logFileName);
+        const editorServicesLogPath = this.logger.getLogFilePath(logFileName);
 
         const psesModulePath =
             path.resolve(
@@ -48,20 +42,20 @@ export class PowerShellProcess {
                 "PowerShellEditorServices/PowerShellEditorServices.psd1");
 
         const featureFlags =
-            this.sessionSettings.developer.featureFlags !== undefined
+            this.sessionSettings.developer.featureFlags.length > 0
                 ? this.sessionSettings.developer.featureFlags.map((f) => `'${f}'`).join(", ")
                 : "";
 
         this.startPsesArgs +=
-            `-LogPath '${PowerShellProcess.escapeSingleQuotes(editorServicesLogPath.fsPath)}' ` +
-        `-SessionDetailsPath '${PowerShellProcess.escapeSingleQuotes(this.sessionFilePath.fsPath)}' ` +
+            `-LogPath '${utils.escapeSingleQuotes(editorServicesLogPath.fsPath)}' ` +
+            `-SessionDetailsPath '${utils.escapeSingleQuotes(this.sessionFilePath.fsPath)}' ` +
             `-FeatureFlags @(${featureFlags}) `;
 
         if (this.sessionSettings.integratedConsole.useLegacyReadLine) {
             this.startPsesArgs += "-UseLegacyReadLine";
         }
 
-        const powerShellArgs = [];
+        const powerShellArgs: string[] = [];
 
         const useLoginShell: boolean =
             (utils.isMacOS && this.sessionSettings.startAsLoginShell.osx)
@@ -80,21 +74,23 @@ export class PowerShellProcess {
         }
 
         const startEditorServices = "Import-Module '" +
-            PowerShellProcess.escapeSingleQuotes(psesModulePath) +
+            utils.escapeSingleQuotes(psesModulePath) +
             "'; Start-EditorServices " + this.startPsesArgs;
 
+        // On Windows we unfortunately can't Base64 encode the startup command
+        // because it annoys some poorly implemented anti-virus scanners.
         if (utils.isWindows) {
             powerShellArgs.push(
                 "-Command",
                 startEditorServices);
         } else {
-            // Use -EncodedCommand for better quote support on non-Windows
+            // Otherwise use -EncodedCommand for better quote support.
             powerShellArgs.push(
                 "-EncodedCommand",
                 Buffer.from(startEditorServices, "utf16le").toString("base64"));
         }
 
-        this.log.write(
+        this.logger.write(
             "Language server starting --",
             "    PowerShell executable: " + this.exePath,
             "    PowerShell args: " + powerShellArgs.join(" "),
@@ -110,58 +106,49 @@ export class PowerShellProcess {
             shellArgs: powerShellArgs,
             cwd: this.sessionSettings.cwd,
             iconPath: new vscode.ThemeIcon("terminal-powershell"),
+            isTransient: true,
+            hideFromUser: this.sessionSettings.integratedConsole.startInBackground,
         };
-
-        if (semver.gte(vscode.version, "1.65.0")) {
-            // @ts-ignore TODO: Don't ignore after we update our engine.
-            terminalOptions.isTransient = true;
-        }
 
         this.consoleTerminal = vscode.window.createTerminal(terminalOptions);
 
         const pwshName = path.basename(this.exePath);
-        this.log.write(`${pwshName} started.`);
+        this.logger.write(`${pwshName} started.`);
 
-        if (this.sessionSettings.integratedConsole.showOnStartup) {
+        if (this.sessionSettings.integratedConsole.showOnStartup
+            && !this.sessionSettings.integratedConsole.startInBackground) {
             // We still need to run this to set the active terminal to the extension terminal.
             this.consoleTerminal.show(true);
         }
 
         // Start the language client
-        this.log.write("Waiting for session file");
         const sessionDetails = await this.waitForSessionFile();
 
         // Subscribe a log event for when the terminal closes
-        this.log.write("Registering terminal close callback");
         this.consoleCloseSubscription = vscode.window.onDidCloseTerminal((terminal) => this.onTerminalClose(terminal));
 
         // Log that the PowerShell terminal process has been started
-        this.log.write("Registering terminal PID log callback");
-        this.consoleTerminal.processId.then((pid) => this.logTerminalPid(pid, pwshName));
+        const pid = await this.consoleTerminal.processId;
+        this.logTerminalPid(pid ?? 0, pwshName);
 
         return sessionDetails;
     }
 
-    public showConsole(preserveFocus: boolean) {
-        if (this.consoleTerminal) {
-            this.consoleTerminal.show(preserveFocus);
-        }
+    public showTerminal(preserveFocus?: boolean) {
+        this.consoleTerminal?.show(preserveFocus);
     }
 
-    public dispose() {
+    public async dispose() {
         // Clean up the session file
-        PowerShellProcess.deleteSessionFile(this.sessionFilePath);
+        this.logger.write("Terminating PowerShell process...");
 
-        if (this.consoleCloseSubscription) {
-            this.consoleCloseSubscription.dispose();
-            this.consoleCloseSubscription = undefined;
-        }
+        await PowerShellProcess.deleteSessionFile(this.sessionFilePath);
 
-        if (this.consoleTerminal) {
-            this.log.write("Terminating PowerShell process...");
-            this.consoleTerminal.dispose();
-            this.consoleTerminal = undefined;
-        }
+        this.consoleCloseSubscription?.dispose();
+        this.consoleCloseSubscription = undefined;
+
+        this.consoleTerminal?.dispose();
+        this.consoleTerminal = undefined;
     }
 
     public sendKeyPress() {
@@ -169,11 +156,11 @@ export class PowerShellProcess {
         // because non-printing characters can cause havoc with different
         // languages and terminal settings. We discard the character server-side
         // anyway, so it doesn't matter what we send.
-        this.consoleTerminal.sendText("p", false);
+        this.consoleTerminal?.sendText("p", false);
     }
 
     private logTerminalPid(pid: number, exeName: string) {
-        this.log.write(`${exeName} PID: ${pid}`);
+        this.logger.write(`${exeName} PID: ${pid}`);
     }
 
     private isLoginShell(pwshPath: string): boolean {
@@ -190,10 +177,9 @@ export class PowerShellProcess {
         return true;
     }
 
-    private static readSessionFile(sessionFilePath: vscode.Uri): IEditorServicesSessionDetails {
-        // TODO: Use vscode.workspace.fs.readFile instead of fs.readFileSync.
-        const fileContents = fs.readFileSync(sessionFilePath.fsPath, "utf-8");
-        return JSON.parse(fileContents);
+    private static async readSessionFile(sessionFilePath: vscode.Uri): Promise<IEditorServicesSessionDetails> {
+        const fileContents = await vscode.workspace.fs.readFile(sessionFilePath);
+        return JSON.parse(fileContents.toString());
     }
 
     private static async deleteSessionFile(sessionFilePath: vscode.Uri) {
@@ -210,25 +196,25 @@ export class PowerShellProcess {
         const warnAt = numOfTries - PowerShellProcess.warnUserThreshold;
 
         // Check every 2 seconds
+        this.logger.write("Waiting for session file...");
         for (let i = numOfTries; i > 0; i--) {
-            if (utils.checkIfFileExists(this.sessionFilePath.fsPath)) {
-                this.log.write("Session file found");
-                const sessionDetails = PowerShellProcess.readSessionFile(this.sessionFilePath);
-                PowerShellProcess.deleteSessionFile(this.sessionFilePath);
+            if (await utils.checkIfFileExists(this.sessionFilePath)) {
+                this.logger.write("Session file found!");
+                const sessionDetails = await PowerShellProcess.readSessionFile(this.sessionFilePath);
+                await PowerShellProcess.deleteSessionFile(this.sessionFilePath);
                 return sessionDetails;
             }
 
             if (warnAt === i) {
-                vscode.window.showWarningMessage(`Loading the PowerShell extension is taking longer than expected.
-    If you're using privilege enforcement software, this can affect start up performance.`);
+                void this.logger.writeAndShowWarning("Loading the PowerShell extension is taking longer than expected. If you're using privilege enforcement software, this can affect start up performance.");
             }
 
             // Wait a bit and try again
             await utils.sleep(2000);
         }
 
-        const err = "Timed out waiting for session file to appear.";
-        this.log.write(err);
+        const err = "Timed out waiting for session file to appear!";
+        this.logger.write(err);
         throw new Error(err);
     }
 
@@ -237,7 +223,7 @@ export class PowerShellProcess {
             return;
         }
 
-        this.log.write("powershell.exe terminated or terminal UI was closed");
+        this.logger.write("powershell.exe terminated or terminal UI was closed");
         this.onExitedEmitter.fire();
     }
 }
