@@ -1,24 +1,67 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-"use strict";
-
 import vscode = require("vscode");
 import utils = require("./utils");
 import os = require("os");
+import { ILogger } from "./logging";
 
-enum CodeFormattingPreset {
-    Custom,
-    Allman,
-    OTBS,
-    Stroustrup,
+// TODO: Quite a few of these settings are unused in the client and instead
+// exist just for the server. Those settings do not need to be represented in
+// this class, as the LSP layers take care of communicating them. Frankly, this
+// class is over-engineered and seems to have originally been created to avoid
+// using vscode.workspace.getConfiguration() directly. It wasn't a bad idea to
+// keep things organized so consistent...but it ended up failing in execution.
+// Perhaps we just get rid of this entirely?
+
+// eslint-disable-next-line @typescript-eslint/no-extraneous-class
+class PartialSettings { }
+
+export class Settings extends PartialSettings {
+    powerShellAdditionalExePaths: PowerShellAdditionalExePathSettings = {};
+    powerShellDefaultVersion = "";
+    // This setting is no longer used but is here to assist in cleaning up the users settings.
+    powerShellExePath = "";
+    promptToUpdatePowerShell = true;
+    startAsLoginShell = new StartAsLoginShellSettings();
+    startAutomatically = true;
+    enableProfileLoading = true;
+    helpCompletion = CommentType.BlockComment;
+    scriptAnalysis = new ScriptAnalysisSettings();
+    debugging = new DebuggingSettings();
+    developer = new DeveloperSettings();
+    codeFormatting = new CodeFormattingSettings();
+    integratedConsole = new IntegratedConsoleSettings();
+    sideBar = new SideBarSettings();
+    pester = new PesterSettings();
+    buttons = new ButtonSettings();
+    cwd = "";
+    enableReferencesCodeLens = true;
+    analyzeOpenDocumentsOnly = false;
+    // TODO: Add (deprecated) useX86Host (for testing)
 }
 
-enum PipelineIndentationStyle {
-    IncreaseIndentationForFirstPipeline,
-    IncreaseIndentationAfterEveryPipeline,
-    NoIndentation,
-    None,
+export enum CodeFormattingPreset {
+    Custom = "Custom",
+    Allman = "Allman",
+    OTBS = "OTBS",
+    Stroustrup = "Stroustrup",
+}
+
+export enum PipelineIndentationStyle {
+    IncreaseIndentationForFirstPipeline = "IncreaseIndentationForFirstPipeline",
+    IncreaseIndentationAfterEveryPipeline = "IncreaseIndentationAfterEveryPipeline",
+    NoIndentation = "NoIndentation",
+    None = "None",
+}
+
+export enum LogLevel {
+    Diagnostic = "Diagnostic",
+    Verbose = "Verbose",
+    Normal = "Normal",
+    Warning = "Warning",
+    Error = "Error",
+    None = "None",
 }
 
 export enum CommentType {
@@ -27,315 +70,179 @@ export enum CommentType {
     LineComment = "LineComment",
 }
 
-export interface IPowerShellAdditionalExePathSettings {
-    [versionName: string]: string;
+export type PowerShellAdditionalExePathSettings = Record<string, string>;
+
+class CodeFormattingSettings extends PartialSettings {
+    autoCorrectAliases = false;
+    avoidSemicolonsAsLineTerminators = false;
+    preset = CodeFormattingPreset.Custom;
+    openBraceOnSameLine = true;
+    newLineAfterOpenBrace = true;
+    newLineAfterCloseBrace = true;
+    pipelineIndentationStyle = PipelineIndentationStyle.NoIndentation;
+    whitespaceBeforeOpenBrace = true;
+    whitespaceBeforeOpenParen = true;
+    whitespaceAroundOperator = true;
+    whitespaceAfterSeparator = true;
+    whitespaceBetweenParameters = false;
+    whitespaceInsideBrace = true;
+    addWhitespaceAroundPipe = true;
+    trimWhitespaceAroundPipe = false;
+    ignoreOneLineBlock = true;
+    alignPropertyValuePairs = true;
+    useConstantStrings = false;
+    useCorrectCasing = false;
 }
 
-export interface IBugReportingSettings {
-    project: string;
+class ScriptAnalysisSettings extends PartialSettings {
+    enable = true;
+    settingsPath = "PSScriptAnalyzerSettings.psd1";
 }
 
-export interface ICodeFoldingSettings {
-    enable?: boolean;
-    showLastLine?: boolean;
+class DebuggingSettings extends PartialSettings {
+    createTemporaryIntegratedConsole = false;
 }
 
-export interface ICodeFormattingSettings {
-    autoCorrectAliases: boolean;
-    preset: CodeFormattingPreset;
-    openBraceOnSameLine: boolean;
-    newLineAfterOpenBrace: boolean;
-    newLineAfterCloseBrace: boolean;
-    pipelineIndentationStyle: PipelineIndentationStyle;
-    whitespaceBeforeOpenBrace: boolean;
-    whitespaceBeforeOpenParen: boolean;
-    whitespaceAroundOperator: boolean;
-    whitespaceAfterSeparator: boolean;
-    whitespaceBetweenParameters: boolean;
-    whitespaceInsideBrace: boolean;
-    addWhitespaceAroundPipe: boolean;
-    trimWhitespaceAroundPipe: boolean;
-    ignoreOneLineBlock: boolean;
-    alignPropertyValuePairs: boolean;
-    useConstantStrings: boolean;
-    useCorrectCasing: boolean;
+class DeveloperSettings extends PartialSettings {
+    featureFlags: string[] = [];
+    // From `<root>/out/main.js` we go to the directory before <root> and
+    // then into the other repo.
+    bundledModulesPath = "../../PowerShellEditorServices/module";
+    editorServicesLogLevel = LogLevel.Normal;
+    editorServicesWaitForDebugger = false;
+    waitForSessionFileTimeoutSeconds = 240;
 }
 
-export interface IScriptAnalysisSettings {
-    enable?: boolean;
-    settingsPath: string;
+// We follow the same convention as VS Code - https://github.com/microsoft/vscode/blob/ff00badd955d6cfcb8eab5f25f3edc86b762f49f/src/vs/workbench/contrib/terminal/browser/terminal.contribution.ts#L105-L107
+//   "Unlike on Linux, ~/.profile is not sourced when logging into a macOS session. This
+//   is the reason terminals on macOS typically run login shells by default which set up
+//   the environment. See http://unix.stackexchange.com/a/119675/115410"
+class StartAsLoginShellSettings extends PartialSettings {
+    osx = true;
+    linux = false;
 }
 
-export interface IDebuggingSettings {
-    createTemporaryIntegratedConsole?: boolean;
+class IntegratedConsoleSettings extends PartialSettings {
+    showOnStartup = true;
+    startInBackground = false;
+    focusConsoleOnExecute = true;
+    useLegacyReadLine = false;
+    forceClearScrollbackBuffer = false;
+    suppressStartupBanner = false;
 }
 
-export interface IDeveloperSettings {
-    featureFlags?: string[];
-    bundledModulesPath?: string;
-    editorServicesLogLevel?: string;
-    editorServicesWaitForDebugger?: boolean;
-    waitForSessionFileTimeoutSeconds?: number;
+class SideBarSettings extends PartialSettings {
+    CommandExplorerVisibility = true;
+    CommandExplorerExcludeFilter: string[] = [];
 }
 
-export interface ISettings {
-    powerShellAdditionalExePaths?: IPowerShellAdditionalExePathSettings;
-    powerShellDefaultVersion?: string;
-    // This setting is no longer used but is here to assist in cleaning up the users settings.
-    powerShellExePath?: string;
-    promptToUpdatePowerShell?: boolean;
-    bundledModulesPath?: string;
-    startAsLoginShell?: IStartAsLoginShellSettings;
-    startAutomatically?: boolean;
-    enableProfileLoading?: boolean;
-    helpCompletion: string;
-    scriptAnalysis?: IScriptAnalysisSettings;
-    debugging?: IDebuggingSettings;
-    developer?: IDeveloperSettings;
-    codeFolding?: ICodeFoldingSettings;
-    codeFormatting?: ICodeFormattingSettings;
-    integratedConsole?: IIntegratedConsoleSettings;
-    bugReporting?: IBugReportingSettings;
-    sideBar?: ISideBarSettings;
-    pester?: IPesterSettings;
-    buttons?: IButtonSettings;
-    cwd?: string;
-    notebooks?: INotebooksSettings;
+class PesterSettings extends PartialSettings {
+    useLegacyCodeLens = true;
+    outputVerbosity = "FromPreference";
+    debugOutputVerbosity = "Diagnostic";
 }
 
-export interface IStartAsLoginShellSettings {
-    osx?: boolean;
-    linux?: boolean;
+class ButtonSettings extends PartialSettings {
+    showRunButtons = true;
+    showPanelMovementButtons = false;
 }
 
-export interface IIntegratedConsoleSettings {
-    showOnStartup?: boolean;
-    focusConsoleOnExecute?: boolean;
-    useLegacyReadLine?: boolean;
-    forceClearScrollbackBuffer?: boolean;
-    suppressStartupBanner?: boolean;
+// This is a recursive function which unpacks a WorkspaceConfiguration into our settings.
+function getSetting<TSetting>(key: string | undefined, value: TSetting, configuration: vscode.WorkspaceConfiguration): TSetting {
+    // Base case where we're looking at a primitive type (or our special record).
+    if (key !== undefined && !(value instanceof PartialSettings)) {
+        return configuration.get<TSetting>(key, value);
+    }
+
+    // Otherwise we're looking at one of our interfaces and need to extract.
+    for (const property in value) {
+        const subKey = key !== undefined ? `${key}.${property}` : property;
+        value[property] = getSetting(subKey, value[property], configuration);
+    }
+
+    return value;
 }
 
-export interface ISideBarSettings {
-    CommandExplorerVisibility?: boolean;
-}
-
-export interface IPesterSettings {
-    useLegacyCodeLens?: boolean;
-    outputVerbosity?: string;
-    debugOutputVerbosity?: string;
-}
-
-export interface IButtonSettings {
-    showRunButtons?: boolean;
-    showPanelMovementButtons?: boolean;
-}
-
-export interface INotebooksSettings {
-    saveMarkdownCellsAs?: CommentType;
-}
-
-// TODO: This could probably be async, and call `validateCwdSetting()` directly.
-export function load(): ISettings {
+export function getSettings(): Settings {
     const configuration: vscode.WorkspaceConfiguration =
         vscode.workspace.getConfiguration(utils.PowerShellLanguageId);
 
-    const defaultBugReportingSettings: IBugReportingSettings = {
-        project: "https://github.com/PowerShell/vscode-powershell",
-    };
-
-    const defaultScriptAnalysisSettings: IScriptAnalysisSettings = {
-        enable: true,
-        settingsPath: "PSScriptAnalyzerSettings.psd1",
-    };
-
-    const defaultDebuggingSettings: IDebuggingSettings = {
-        createTemporaryIntegratedConsole: false,
-    };
-
-    const defaultDeveloperSettings: IDeveloperSettings = {
-        featureFlags: [],
-        // From `<root>/out/main.js` we go to the directory before <root> and
-        // then into the other repo.
-        bundledModulesPath: "../../PowerShellEditorServices/module",
-        editorServicesLogLevel: "Normal",
-        editorServicesWaitForDebugger: false,
-        waitForSessionFileTimeoutSeconds: 240,
-    };
-
-    const defaultCodeFoldingSettings: ICodeFoldingSettings = {
-        enable: true,
-        showLastLine: false,
-    };
-
-    const defaultCodeFormattingSettings: ICodeFormattingSettings = {
-        autoCorrectAliases: false,
-        preset: CodeFormattingPreset.Custom,
-        openBraceOnSameLine: true,
-        newLineAfterOpenBrace: true,
-        newLineAfterCloseBrace: true,
-        pipelineIndentationStyle: PipelineIndentationStyle.NoIndentation,
-        whitespaceBeforeOpenBrace: true,
-        whitespaceBeforeOpenParen: true,
-        whitespaceAroundOperator: true,
-        whitespaceAfterSeparator: true,
-        whitespaceBetweenParameters: false,
-        whitespaceInsideBrace: true,
-        addWhitespaceAroundPipe: true,
-        trimWhitespaceAroundPipe: false,
-        ignoreOneLineBlock: true,
-        alignPropertyValuePairs: true,
-        useConstantStrings: false,
-        useCorrectCasing: false,
-    };
-
-    const defaultStartAsLoginShellSettings: IStartAsLoginShellSettings = {
-        osx: true,
-        linux: false,
-    };
-
-    const defaultIntegratedConsoleSettings: IIntegratedConsoleSettings = {
-        showOnStartup: true,
-        focusConsoleOnExecute: true,
-        useLegacyReadLine: false,
-        forceClearScrollbackBuffer: false,
-    };
-
-    const defaultSideBarSettings: ISideBarSettings = {
-        CommandExplorerVisibility: true,
-    };
-
-    const defaultButtonSettings: IButtonSettings = {
-        showRunButtons: true,
-        showPanelMovementButtons: false
-    };
-
-    const defaultPesterSettings: IPesterSettings = {
-        useLegacyCodeLens: true,
-        outputVerbosity: "FromPreference",
-        debugOutputVerbosity: "Diagnostic",
-    };
-
-    const defaultNotebooksSettings: INotebooksSettings = {
-        saveMarkdownCellsAs: CommentType.BlockComment,
-    };
-
-    return {
-        startAutomatically:
-            configuration.get<boolean>("startAutomatically", true),
-        powerShellAdditionalExePaths:
-            configuration.get<IPowerShellAdditionalExePathSettings>("powerShellAdditionalExePaths", undefined),
-        powerShellDefaultVersion:
-            configuration.get<string>("powerShellDefaultVersion", undefined),
-        powerShellExePath:
-            configuration.get<string>("powerShellExePath", undefined),
-        promptToUpdatePowerShell:
-            configuration.get<boolean>("promptToUpdatePowerShell", true),
-        bundledModulesPath:
-            "../modules", // Because the extension is always at `<root>/out/main.js`
-        enableProfileLoading:
-            configuration.get<boolean>("enableProfileLoading", false),
-        helpCompletion:
-            configuration.get<string>("helpCompletion", CommentType.BlockComment),
-        scriptAnalysis:
-            configuration.get<IScriptAnalysisSettings>("scriptAnalysis", defaultScriptAnalysisSettings),
-        debugging:
-            configuration.get<IDebuggingSettings>("debugging", defaultDebuggingSettings),
-        developer:
-            getWorkspaceSettingsWithDefaults<IDeveloperSettings>(configuration, "developer", defaultDeveloperSettings),
-        codeFolding:
-            configuration.get<ICodeFoldingSettings>("codeFolding", defaultCodeFoldingSettings),
-        codeFormatting:
-            configuration.get<ICodeFormattingSettings>("codeFormatting", defaultCodeFormattingSettings),
-        integratedConsole:
-            configuration.get<IIntegratedConsoleSettings>("integratedConsole", defaultIntegratedConsoleSettings),
-        bugReporting:
-            configuration.get<IBugReportingSettings>("bugReporting", defaultBugReportingSettings),
-        sideBar:
-            configuration.get<ISideBarSettings>("sideBar", defaultSideBarSettings),
-        pester:
-            configuration.get<IPesterSettings>("pester", defaultPesterSettings),
-        buttons:
-            configuration.get<IButtonSettings>("buttons", defaultButtonSettings),
-        notebooks:
-            configuration.get<INotebooksSettings>("notebooks", defaultNotebooksSettings),
-        startAsLoginShell:
-            // tslint:disable-next-line
-            // We follow the same convention as VS Code - https://github.com/microsoft/vscode/blob/ff00badd955d6cfcb8eab5f25f3edc86b762f49f/src/vs/workbench/contrib/terminal/browser/terminal.contribution.ts#L105-L107
-            //   "Unlike on Linux, ~/.profile is not sourced when logging into a macOS session. This
-            //   is the reason terminals on macOS typically run login shells by default which set up
-            //   the environment. See http://unix.stackexchange.com/a/119675/115410"
-            configuration.get<IStartAsLoginShellSettings>("startAsLoginShell", defaultStartAsLoginShellSettings),
-        cwd: // NOTE: This must be validated at startup via `validateCwdSetting()`. There's probably a better way to do this.
-            configuration.get<string>("cwd", undefined),
-    };
+    return getSetting(undefined, new Settings(), configuration);
 }
 
 // Get the ConfigurationTarget (read: scope) of where the *effective* setting value comes from
-export async function getEffectiveConfigurationTarget(settingName: string): Promise<vscode.ConfigurationTarget> {
+export function getEffectiveConfigurationTarget(settingName: string): vscode.ConfigurationTarget | undefined {
     const configuration = vscode.workspace.getConfiguration(utils.PowerShellLanguageId);
     const detail = configuration.inspect(settingName);
-    let configurationTarget = null;
-
-    if (typeof detail.workspaceFolderValue !== "undefined") {
-        configurationTarget = vscode.ConfigurationTarget.WorkspaceFolder;
+    if (detail === undefined) {
+        return undefined;
+    } else if (typeof detail.workspaceFolderValue !== "undefined") {
+        return vscode.ConfigurationTarget.WorkspaceFolder;
     }
     else if (typeof detail.workspaceValue !== "undefined") {
-        configurationTarget = vscode.ConfigurationTarget.Workspace;
+        return vscode.ConfigurationTarget.Workspace;
     }
     else if (typeof detail.globalValue !== "undefined") {
-        configurationTarget = vscode.ConfigurationTarget.Global;
+        return vscode.ConfigurationTarget.Global;
     }
-    return configurationTarget;
+    return undefined;
 }
 
-export async function change(
+export async function changeSetting(
     settingName: string,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     newValue: any,
-    configurationTarget?: vscode.ConfigurationTarget | boolean): Promise<void> {
+    configurationTarget: vscode.ConfigurationTarget | boolean | undefined,
+    logger: ILogger | undefined): Promise<void> {
 
-    const configuration = vscode.workspace.getConfiguration(utils.PowerShellLanguageId);
-    await configuration.update(settingName, newValue, configurationTarget);
-}
+    logger?.writeDiagnostic(`Changing '${settingName}' at scope '${configurationTarget} to '${newValue}'`);
 
-function getWorkspaceSettingsWithDefaults<TSettings>(
-    workspaceConfiguration: vscode.WorkspaceConfiguration,
-    settingName: string,
-    defaultSettings: TSettings): TSettings {
-
-    const importedSettings: TSettings = workspaceConfiguration.get<TSettings>(settingName, defaultSettings);
-
-    for (const setting in importedSettings) {
-        if (importedSettings[setting]) {
-            defaultSettings[setting] = importedSettings[setting];
-        }
+    try {
+        const configuration = vscode.workspace.getConfiguration(utils.PowerShellLanguageId);
+        await configuration.update(settingName, newValue, configurationTarget);
+    } catch (err) {
+        logger?.writeError(`Failed to change setting: ${err}`);
     }
-    return defaultSettings;
 }
 
-export async function validateCwdSetting(): Promise<string> {
-    let cwd: string = vscode.workspace.getConfiguration(utils.PowerShellLanguageId).get<string>("cwd", null);
+// We don't want to query the user more than once, so this is idempotent.
+let hasPrompted = false;
+export let chosenWorkspace: vscode.WorkspaceFolder | undefined = undefined;
+
+export async function validateCwdSetting(logger: ILogger): Promise<string> {
+    let cwd: string | undefined = vscode.workspace.getConfiguration(utils.PowerShellLanguageId).get<string>("cwd");
 
     // Only use the cwd setting if it exists.
-    if (utils.checkIfDirectoryExists(cwd)) {
-        return cwd;
-    } else {
-        // Otherwise use a workspace folder, prompting if necessary.
-        if (vscode.workspace.workspaceFolders?.length > 1) {
-            const options: vscode.WorkspaceFolderPickOptions = {
-                placeHolder: "Select a folder to use as the PowerShell extension's working directory.",
-            }
-            cwd = (await vscode.window.showWorkspaceFolderPick(options))?.uri.fsPath;
-            // Save the picked 'cwd' to the workspace settings.
-            await change("cwd", cwd);
-        } else {
-            cwd = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-        }
-        // If there were no workspace folders, or somehow they don't exist, use
-        // the home directory.
-        if (cwd === undefined || !utils.checkIfDirectoryExists(cwd)) {
-            return os.homedir();
-        }
+    if (cwd !== undefined && await utils.checkIfDirectoryExists(cwd)) {
         return cwd;
     }
+
+    // If there is no workspace, or there is but it has no folders, fallback.
+    if (vscode.workspace.workspaceFolders === undefined
+        || vscode.workspace.workspaceFolders.length === 0) {
+        cwd = undefined;
+        // If there is exactly one workspace folder, use that.
+    } else if (vscode.workspace.workspaceFolders.length === 1) {
+        cwd = vscode.workspace.workspaceFolders[0].uri.fsPath;
+        // If there is more than one workspace folder, prompt the user once.
+    } else if (vscode.workspace.workspaceFolders.length > 1 && !hasPrompted) {
+        hasPrompted = true;
+        const options: vscode.WorkspaceFolderPickOptions = {
+            placeHolder: "Select a workspace folder to use for the PowerShell Extension.",
+        };
+        chosenWorkspace = await vscode.window.showWorkspaceFolderPick(options);
+        cwd = chosenWorkspace?.uri.fsPath;
+        // Save the picked 'cwd' to the workspace settings.
+        // We have to check again because the user may not have picked.
+        if (cwd !== undefined && await utils.checkIfDirectoryExists(cwd)) {
+            await changeSetting("cwd", cwd, undefined, logger);
+        }
+    }
+
+    // If there were no workspace folders, or somehow they don't exist, use
+    // the home directory.
+    if (cwd === undefined || !await utils.checkIfDirectoryExists(cwd)) {
+        return os.homedir();
+    }
+    return cwd;
 }
