@@ -1,30 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-"use strict";
-
-import fs = require("fs");
 import os = require("os");
 import path = require("path");
 import vscode = require("vscode");
 
 export const PowerShellLanguageId = "powershell";
 
-// Check that the file exists in an asynchronous manner that relies solely on the VS Code API, not Node's fs library.
-export async function fileExists(targetPath: string | vscode.Uri): Promise<boolean> {
-    try {
-            await vscode.workspace.fs.stat(
-                targetPath instanceof vscode.Uri
-                    ? targetPath
-                    : vscode.Uri.file(targetPath));
-        return true;
-    } catch (e) {
-        if (e instanceof vscode.FileSystemError.FileNotFound) {
-            return false;
-        }
-        throw e;
-    }
-
+export function escapeSingleQuotes(p: string): string {
+    return p.replace(new RegExp("'", "g"), "''");
 }
 
 export function getPipePath(pipeName: string) {
@@ -37,23 +21,38 @@ export function getPipePath(pipeName: string) {
     }
 }
 
-export function checkIfFileExists(filePath: string): boolean {
+// Check that the file or directory exists in an asynchronous manner that relies
+// solely on the VS Code API, not Node's fs library, ignoring symlinks.
+async function checkIfFileOrDirectoryExists(targetPath: string | vscode.Uri, type: vscode.FileType): Promise<boolean> {
+    if (targetPath === "") {
+        return false;
+    }
     try {
-        fs.accessSync(filePath, fs.constants.R_OK);
-        return true;
-    } catch (e) {
+        const stat: vscode.FileStat = await vscode.workspace.fs.stat(
+            targetPath instanceof vscode.Uri
+                ? targetPath
+                : vscode.Uri.file(targetPath));
+        return (stat.type & type) !== 0;
+    } catch {
+        // TODO: Maybe throw if it's not a FileNotFound exception.
         return false;
     }
 }
 
-export function checkIfDirectoryExists(directoryPath: string): boolean {
-    try {
-        // tslint:disable-next-line:no-bitwise
-        fs.accessSync(directoryPath, fs.constants.R_OK | fs.constants.O_DIRECTORY);
-        return true;
-    } catch (e) {
-        return false;
-    }
+export async function checkIfFileExists(filePath: string | vscode.Uri): Promise<boolean> {
+    return await checkIfFileOrDirectoryExists(filePath, vscode.FileType.File);
+}
+
+export async function checkIfDirectoryExists(directoryPath: string | vscode.Uri): Promise<boolean> {
+    return await checkIfFileOrDirectoryExists(directoryPath, vscode.FileType.Directory);
+}
+
+export async function readDirectory(directoryPath: string | vscode.Uri): Promise<string[]> {
+    const items = await vscode.workspace.fs.readDirectory(
+        directoryPath instanceof vscode.Uri
+            ? directoryPath
+            : vscode.Uri.file(directoryPath));
+    return items.map(([name, _type]) => name);
 }
 
 export function getTimestampString() {
