@@ -5,8 +5,38 @@ import * as assert from "assert";
 import mockFS = require("mock-fs");
 import FileSystem = require("mock-fs/lib/filesystem");
 import * as path from "path";
+import rewire = require("rewire");
 import * as sinon from "sinon";
 import * as platform from "../../src/platform";
+import * as fs from "fs"; // NOTE: Necessary for mock-fs.
+import * as vscode from "vscode";
+
+// We have to rewire the platform module so that mock-fs can be used, as it
+// overrides the fs module but not the vscode.workspace.fs module.
+const platformMock = rewire("../../src/platform");
+
+// eslint-disable-next-line @typescript-eslint/require-await
+async function fakeCheckIfFileOrDirectoryExists(targetPath: string | vscode.Uri): Promise<boolean> {
+    try {
+        fs.lstatSync(targetPath instanceof vscode.Uri ? targetPath.fsPath : targetPath);
+        return true;
+    } catch {
+        return false;
+    }
+}
+
+// eslint-disable-next-line @typescript-eslint/require-await
+async function fakeReadDirectory(targetPath: string | vscode.Uri): Promise<string[]> {
+    return fs.readdirSync(targetPath instanceof vscode.Uri ? targetPath.fsPath : targetPath);
+}
+
+const utilsMock = {
+    checkIfFileExists: fakeCheckIfFileOrDirectoryExists,
+    checkIfDirectoryExists: fakeCheckIfFileOrDirectoryExists,
+    readDirectory: fakeReadDirectory
+};
+
+platformMock.__set__("utils", utilsMock);
 
 /**
  * Describes a platform on which the PowerShell extension should work,
@@ -16,7 +46,7 @@ interface ITestPlatform {
     name: string;
     platformDetails: platform.IPlatformDetails;
     filesystem: FileSystem.DirectoryItems;
-    environmentVars?: Record<string, string>;
+    environmentVars: Record<string, string>;
 }
 
 /**
@@ -31,11 +61,12 @@ interface ITestPlatformSuccessCase extends ITestPlatform {
 // Platform configurations where we expect to find a set of PowerShells
 let successTestCases: ITestPlatformSuccessCase[];
 
-let msixAppDir = null;
-let pwshMsixPath = null;
-let pwshPreviewMsixPath = null;
+let msixAppDir: string;
+let pwshMsixPath: string;
+let pwshPreviewMsixPath: string;
+
 if (process.platform === "win32") {
-    msixAppDir = path.join(process.env.LOCALAPPDATA, "Microsoft", "WindowsApps");
+    msixAppDir = path.join(process.env.LOCALAPPDATA!, "Microsoft", "WindowsApps");
     pwshMsixPath = path.join(msixAppDir, "Microsoft.PowerShell_8wekyb3d8bbwe", "pwsh.exe");
     pwshPreviewMsixPath = path.join(msixAppDir, "Microsoft.PowerShellPreview_8wekyb3d8bbwe", "pwsh.exe");
 
@@ -56,34 +87,42 @@ if (process.platform === "win32") {
                 {
                     exePath: "C:\\Program Files\\PowerShell\\6\\pwsh.exe",
                     displayName: "PowerShell (x64)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\Program Files (x86)\\PowerShell\\6\\pwsh.exe",
                     displayName: "PowerShell (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: pwshMsixPath,
                     displayName: "PowerShell (Store)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\Program Files\\PowerShell\\7-preview\\pwsh.exe",
                     displayName: "PowerShell Preview (x64)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: pwshPreviewMsixPath,
                     displayName: "PowerShell Preview (Store)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\Program Files (x86)\\PowerShell\\7-preview\\pwsh.exe",
                     displayName: "PowerShell Preview (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x64)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\WINDOWS\\SysWOW64\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x86)",
+                    supportsProperArguments: true
                 },
             ],
             filesystem: {
@@ -135,10 +174,12 @@ if (process.platform === "win32") {
                 {
                     exePath: "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x64)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\WINDOWS\\SysWOW64\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x86)",
+                    supportsProperArguments: true
                 },
             ],
             filesystem: {
@@ -166,34 +207,42 @@ if (process.platform === "win32") {
                 {
                     exePath: "C:\\Program Files (x86)\\PowerShell\\6\\pwsh.exe",
                     displayName: "PowerShell (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\Program Files\\PowerShell\\6\\pwsh.exe",
                     displayName: "PowerShell (x64)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: pwshMsixPath,
                     displayName: "PowerShell (Store)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\Program Files (x86)\\PowerShell\\7-preview\\pwsh.exe",
                     displayName: "PowerShell Preview (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: pwshPreviewMsixPath,
                     displayName: "PowerShell Preview (Store)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\Program Files\\PowerShell\\7-preview\\pwsh.exe",
                     displayName: "PowerShell Preview (x64)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\WINDOWS\\Sysnative\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x64)",
+                    supportsProperArguments: true
                 },
             ],
             filesystem: {
@@ -245,10 +294,12 @@ if (process.platform === "win32") {
                 {
                     exePath: "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\WINDOWS\\Sysnative\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x64)",
+                    supportsProperArguments: true
                 },
             ],
             filesystem: {
@@ -276,22 +327,27 @@ if (process.platform === "win32") {
                 {
                     exePath: "C:\\Program Files (x86)\\PowerShell\\6\\pwsh.exe",
                     displayName: "PowerShell (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: pwshMsixPath,
                     displayName: "PowerShell (Store)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\Program Files (x86)\\PowerShell\\7-preview\\pwsh.exe",
                     displayName: "PowerShell Preview (x86)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: pwshPreviewMsixPath,
                     displayName: "PowerShell Preview (Store)",
+                    supportsProperArguments: true
                 },
                 {
                     exePath: "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x86)",
+                    supportsProperArguments: true
                 },
             ],
             filesystem: {
@@ -332,11 +388,49 @@ if (process.platform === "win32") {
                 {
                     exePath: "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
                     displayName: "Windows PowerShell (x86)",
+                    supportsProperArguments: true
                 },
             ],
             filesystem: {
                 "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0": {
                     "powershell.exe": "",
+                },
+            },
+        },
+        {
+            name: "Windows (dotnet)",
+            platformDetails: {
+                operatingSystem: platform.OperatingSystem.Windows,
+                isOS64Bit: true,
+                isProcess64Bit: true,
+            },
+            environmentVars: {
+                "USERNAME": "test",
+                "USERPROFILE": "C:\\Users\\test",
+                "ProgramFiles": "C:\\Program Files",
+                "ProgramFiles(x86)": "C:\\Program Files (x86)",
+                "windir": "C:\\WINDOWS",
+            },
+            expectedPowerShellSequence: [
+                {
+                    exePath: "C:\\Users\\test\\.dotnet\\tools\\pwsh.exe",
+                    displayName: ".NET Core PowerShell Global Tool",
+                    supportsProperArguments: false
+                },
+                {
+                    exePath: "C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
+                    displayName: "Windows PowerShell (x64)",
+                    supportsProperArguments: true
+                },
+                {
+                    exePath: "C:\\WINDOWS\\SysWOW64\\WindowsPowerShell\\v1.0\\powershell.exe",
+                    displayName: "Windows PowerShell (x86)",
+                    supportsProperArguments: true
+                },
+            ],
+            filesystem: {
+                "C:\\Users\\test\\.dotnet\\tools": {
+                    "pwsh.exe": "",
                 },
             },
         },
@@ -350,11 +444,28 @@ if (process.platform === "win32") {
                 isOS64Bit: true,
                 isProcess64Bit: true,
             },
+            environmentVars: {},
             expectedPowerShellSequence: [
-                { exePath: "/usr/bin/pwsh", displayName: "PowerShell" },
-                { exePath: "/snap/bin/pwsh", displayName: "PowerShell Snap" },
-                { exePath: "/usr/bin/pwsh-preview", displayName: "PowerShell Preview" },
-                { exePath: "/snap/bin/pwsh-preview", displayName: "PowerShell Preview Snap" },
+                {
+                    exePath: "/usr/bin/pwsh",
+                    displayName: "PowerShell",
+                    supportsProperArguments: true
+                },
+                {
+                    exePath: "/snap/bin/pwsh",
+                    displayName: "PowerShell Snap",
+                    supportsProperArguments: true
+                },
+                {
+                    exePath: "/usr/bin/pwsh-preview",
+                    displayName: "PowerShell Preview",
+                    supportsProperArguments: true
+                },
+                {
+                    exePath: "/snap/bin/pwsh-preview",
+                    displayName: "PowerShell Preview Snap",
+                    supportsProperArguments: true
+                },
             ],
             filesystem: {
                 "/usr/bin": {
@@ -374,9 +485,18 @@ if (process.platform === "win32") {
                 isOS64Bit: true,
                 isProcess64Bit: true,
             },
+            environmentVars: {},
             expectedPowerShellSequence: [
-                { exePath: "/usr/local/bin/pwsh", displayName: "PowerShell" },
-                { exePath: "/usr/local/bin/pwsh-preview", displayName: "PowerShell Preview" },
+                {
+                    exePath: "/usr/local/bin/pwsh",
+                    displayName: "PowerShell",
+                    supportsProperArguments: true
+                },
+                {
+                    exePath: "/usr/local/bin/pwsh-preview",
+                    displayName: "PowerShell Preview",
+                    supportsProperArguments: true
+                },
             ],
             filesystem: {
                 "/usr/local/bin": {
@@ -392,8 +512,13 @@ if (process.platform === "win32") {
                 isOS64Bit: true,
                 isProcess64Bit: true,
             },
+            environmentVars: {},
             expectedPowerShellSequence: [
-                { exePath: "/usr/bin/pwsh", displayName: "PowerShell" },
+                {
+                    exePath: "/usr/bin/pwsh",
+                    displayName: "PowerShell",
+                    supportsProperArguments: true
+                },
             ],
             filesystem: {
                 "/usr/bin": {
@@ -408,8 +533,13 @@ if (process.platform === "win32") {
                 isOS64Bit: true,
                 isProcess64Bit: true,
             },
+            environmentVars: {},
             expectedPowerShellSequence: [
-                { exePath: "/snap/bin/pwsh", displayName: "PowerShell Snap" },
+                {
+                    exePath: "/snap/bin/pwsh",
+                    displayName: "PowerShell Snap",
+                    supportsProperArguments: true
+                },
             ],
             filesystem: {
                 "/snap/bin": {
@@ -424,11 +554,64 @@ if (process.platform === "win32") {
                 isOS64Bit: true,
                 isProcess64Bit: true,
             },
+            environmentVars: {},
             expectedPowerShellSequence: [
-                { exePath: "/usr/local/bin/pwsh", displayName: "PowerShell" },
+                {
+                    exePath: "/usr/local/bin/pwsh",
+                    displayName: "PowerShell",
+                    supportsProperArguments: true
+                },
             ],
             filesystem: {
                 "/usr/local/bin": {
+                    pwsh: "",
+                },
+            },
+        },
+        {
+            name: "MacOS (dotnet)",
+            platformDetails: {
+                operatingSystem: platform.OperatingSystem.MacOS,
+                isOS64Bit: true,
+                isProcess64Bit: true,
+            },
+            environmentVars: {
+                "USER": "test",
+                "HOME": "/Users/test",
+            },
+            expectedPowerShellSequence: [
+                {
+                    exePath: "/Users/test/.dotnet/tools/pwsh",
+                    displayName: ".NET Core PowerShell Global Tool",
+                    supportsProperArguments: false
+                },
+            ],
+            filesystem: {
+                "/Users/test/.dotnet/tools": {
+                    pwsh: "",
+                },
+            },
+        },
+        {
+            name: "Linux (dotnet)",
+            platformDetails: {
+                operatingSystem: platform.OperatingSystem.Linux,
+                isOS64Bit: true,
+                isProcess64Bit: true,
+            },
+            environmentVars: {
+                "USER": "test",
+                "HOME": "/home/test",
+            },
+            expectedPowerShellSequence: [
+                {
+                    exePath: "/home/test/.dotnet/tools/pwsh",
+                    displayName: ".NET Core PowerShell Global Tool",
+                    supportsProperArguments: false
+                },
+            ],
+            filesystem: {
+                "/home/test/.dotnet/tools": {
                     pwsh: "",
                 },
             },
@@ -444,6 +627,7 @@ const errorTestCases: ITestPlatform[] = [
             isOS64Bit: true,
             isProcess64Bit: true,
         },
+        environmentVars: {},
         filesystem: {},
     },
     {
@@ -453,6 +637,7 @@ const errorTestCases: ITestPlatform[] = [
             isOS64Bit: true,
             isProcess64Bit: true,
         },
+        environmentVars: {},
         filesystem: {},
     },
 ];
@@ -460,64 +645,62 @@ const errorTestCases: ITestPlatform[] = [
 function setupTestEnvironment(testPlatform: ITestPlatform) {
     mockFS(testPlatform.filesystem);
 
-    if (testPlatform.environmentVars) {
-        for (const envVar of Object.keys(testPlatform.environmentVars)) {
-            sinon.stub(process.env, envVar).value(testPlatform.environmentVars[envVar]);
-        }
+    for (const envVar of Object.keys(testPlatform.environmentVars)) {
+        sinon.stub(process.env, envVar).value(testPlatform.environmentVars[envVar]);
     }
 }
 
 describe("Platform module", function () {
     it("Gets the correct platform details", function () {
-        const platformDetails: platform.IPlatformDetails = platform.getPlatformDetails();
+        const platformDetails: platform.IPlatformDetails = platformMock.getPlatformDetails();
         switch (process.platform) {
-            case "darwin":
-                assert.strictEqual(
-                    platformDetails.operatingSystem,
-                    platform.OperatingSystem.MacOS,
-                    "Platform details operating system should be MacOS");
-                assert.strictEqual(
-                    platformDetails.isProcess64Bit,
-                    true,
-                    "VSCode on darwin should be 64-bit");
-                assert.strictEqual(
-                    platformDetails.isOS64Bit,
-                    true,
-                    "Darwin is 64-bit only");
-                break;
+        case "darwin":
+            assert.strictEqual(
+                platformDetails.operatingSystem,
+                platform.OperatingSystem.MacOS,
+                "Platform details operating system should be MacOS");
+            assert.strictEqual(
+                platformDetails.isProcess64Bit,
+                true,
+                "VSCode on darwin should be 64-bit");
+            assert.strictEqual(
+                platformDetails.isOS64Bit,
+                true,
+                "Darwin is 64-bit only");
+            break;
 
-            case "linux":
-                assert.strictEqual(
-                    platformDetails.operatingSystem,
-                    platform.OperatingSystem.Linux,
-                    "Platform details operating system should be Linux");
-                assert.strictEqual(
-                    platformDetails.isProcess64Bit,
-                    true,
-                    "Only 64-bit VSCode supported on Linux");
-                assert.strictEqual(
-                    platformDetails.isOS64Bit,
-                    true,
-                    "Only 64-bit Linux supported by PowerShell");
-                return;
+        case "linux":
+            assert.strictEqual(
+                platformDetails.operatingSystem,
+                platform.OperatingSystem.Linux,
+                "Platform details operating system should be Linux");
+            assert.strictEqual(
+                platformDetails.isProcess64Bit,
+                true,
+                "Only 64-bit VSCode supported on Linux");
+            assert.strictEqual(
+                platformDetails.isOS64Bit,
+                true,
+                "Only 64-bit Linux supported by PowerShell");
+            return;
 
-            case "win32":
-                assert.strictEqual(
-                    platformDetails.operatingSystem,
-                    platform.OperatingSystem.Windows,
-                    "Platform details operating system should be Windows");
-                assert.strictEqual(
-                    platformDetails.isProcess64Bit,
-                    process.arch === "x64",
-                    "Windows process bitness should match process arch");
-                assert.strictEqual(
-                    platformDetails.isOS64Bit,
-                    !!(platformDetails.isProcess64Bit || process.env.ProgramW6432),
-                    "Windows OS arch should match process bitness unless 64-bit env var set");
-                return;
+        case "win32":
+            assert.strictEqual(
+                platformDetails.operatingSystem,
+                platform.OperatingSystem.Windows,
+                "Platform details operating system should be Windows");
+            assert.strictEqual(
+                platformDetails.isProcess64Bit,
+                process.arch === "x64",
+                "Windows process bitness should match process arch");
+            assert.strictEqual(
+                platformDetails.isOS64Bit,
+                !!(platformDetails.isProcess64Bit || process.env.ProgramW6432),
+                "Windows OS arch should match process bitness unless 64-bit env var set");
+            return;
 
-            default:
-                assert.fail("This platform is unsupported");
+        default:
+            assert.fail("This platform is unsupported");
         }
     });
 
@@ -528,26 +711,27 @@ describe("Platform module", function () {
         });
 
         for (const testPlatform of successTestCases) {
-            it(`Finds it on ${testPlatform.name}`, function () {
+            it(`Finds it on ${testPlatform.name}`, async function () {
                 setupTestEnvironment(testPlatform);
 
-                const powerShellExeFinder = new platform.PowerShellExeFinder(testPlatform.platformDetails);
+                const powerShellExeFinder = new platformMock.PowerShellExeFinder(testPlatform.platformDetails);
 
-                const defaultPowerShell = powerShellExeFinder.getFirstAvailablePowerShellInstallation();
+                const defaultPowerShell = await powerShellExeFinder.getFirstAvailablePowerShellInstallation();
                 const expectedPowerShell = testPlatform.expectedPowerShellSequence[0];
 
                 assert.strictEqual(defaultPowerShell.exePath, expectedPowerShell.exePath);
                 assert.strictEqual(defaultPowerShell.displayName, expectedPowerShell.displayName);
+                assert.strictEqual(defaultPowerShell.supportsProperArguments, expectedPowerShell.supportsProperArguments);
             });
         }
 
         for (const testPlatform of errorTestCases) {
-            it(`Fails gracefully on ${testPlatform.name}`, function () {
+            it(`Fails gracefully on ${testPlatform.name}`, async function () {
                 setupTestEnvironment(testPlatform);
 
-                const powerShellExeFinder = new platform.PowerShellExeFinder(testPlatform.platformDetails);
+                const powerShellExeFinder = new platformMock.PowerShellExeFinder(testPlatform.platformDetails);
 
-                const defaultPowerShell = powerShellExeFinder.getFirstAvailablePowerShellInstallation();
+                const defaultPowerShell = await powerShellExeFinder.getFirstAvailablePowerShellInstallation();
                 assert.strictEqual(defaultPowerShell, undefined);
             });
         }
@@ -560,19 +744,20 @@ describe("Platform module", function () {
         });
 
         for (const testPlatform of successTestCases) {
-            it(`Finds them on ${testPlatform.name}`, function () {
+            it(`Finds them on ${testPlatform.name}`, async function () {
                 setupTestEnvironment(testPlatform);
 
-                const powerShellExeFinder = new platform.PowerShellExeFinder(testPlatform.platformDetails);
+                const powerShellExeFinder = new platformMock.PowerShellExeFinder(testPlatform.platformDetails);
 
-                const foundPowerShells = powerShellExeFinder.getAllAvailablePowerShellInstallations();
+                const foundPowerShells = await powerShellExeFinder.getAllAvailablePowerShellInstallations();
 
                 for (let i = 0; i < testPlatform.expectedPowerShellSequence.length; i++) {
                     const foundPowerShell = foundPowerShells[i];
                     const expectedPowerShell = testPlatform.expectedPowerShellSequence[i];
 
-                    assert.strictEqual(foundPowerShell && foundPowerShell.exePath, expectedPowerShell.exePath);
-                    assert.strictEqual(foundPowerShell && foundPowerShell.displayName, expectedPowerShell.displayName);
+                    assert.strictEqual(foundPowerShell?.exePath, expectedPowerShell.exePath);
+                    assert.strictEqual(foundPowerShell?.displayName, expectedPowerShell.displayName);
+                    assert.strictEqual(foundPowerShell?.supportsProperArguments, expectedPowerShell.supportsProperArguments);
                 }
 
                 assert.strictEqual(
@@ -583,12 +768,12 @@ describe("Platform module", function () {
         }
 
         for (const testPlatform of errorTestCases) {
-            it(`Fails gracefully on ${testPlatform.name}`, function () {
+            it(`Fails gracefully on ${testPlatform.name}`, async function () {
                 setupTestEnvironment(testPlatform);
 
-                const powerShellExeFinder = new platform.PowerShellExeFinder(testPlatform.platformDetails);
+                const powerShellExeFinder = new platformMock.PowerShellExeFinder(testPlatform.platformDetails);
 
-                const foundPowerShells = powerShellExeFinder.getAllAvailablePowerShellInstallations();
+                const foundPowerShells = await powerShellExeFinder.getAllAvailablePowerShellInstallations();
                 assert.strictEqual(foundPowerShells.length, 0);
             });
         }
@@ -608,7 +793,7 @@ describe("Platform module", function () {
 
                 function getWinPSPath(systemDir: string) {
                     return path.join(
-                        testPlatform.environmentVars.windir,
+                        testPlatform.environmentVars.windir!,
                         systemDir,
                         "WindowsPowerShell",
                         "v1.0",
@@ -626,7 +811,7 @@ describe("Platform module", function () {
                     altWinPSPath = null;
                 }
 
-                const powerShellExeFinder = new platform.PowerShellExeFinder(testPlatform.platformDetails);
+                const powerShellExeFinder = new platformMock.PowerShellExeFinder(testPlatform.platformDetails);
 
                 assert.strictEqual(powerShellExeFinder.fixWindowsPowerShellPath(winPSPath), winPSPath);
 
