@@ -31,11 +31,10 @@ describe("ExternalApi feature", function () {
         });
 
         it("Rejects if not registered", async function () {
-            assert.rejects(
-                async () => await extension.getPowerShellVersionDetails(""))
+            await assert.rejects(async () => await extension.getPowerShellVersionDetails(""));
         });
 
-        it("Throws if attempting to register an extension more than once", async function () {
+        it("Throws if attempting to register an extension more than once", function () {
             const sessionId: string = extension.registerExternalExtension(utils.extensionId);
             try {
                 assert.throws(
@@ -48,11 +47,11 @@ describe("ExternalApi feature", function () {
             }
         });
 
-        it("Throws when unregistering an extension that isn't registered", async function () {
+        it("Throws when unregistering an extension that isn't registered", function () {
             assert.throws(
                 () => extension.unregisterExternalExtension("not-real"),
                 {
-                    message: `No extension registered with session UUID: not-real`
+                    message: "No extension registered with session UUID: not-real"
                 });
         });
     });
