@@ -2,12 +2,11 @@
 // Licensed under the MIT License.
 
 import * as assert from "assert";
-import * as fs from "fs";
 import * as path from "path";
 import rewire = require("rewire");
 import vscode = require("vscode");
 import utils = require("../utils");
-import { sleep } from "../../src/utils";
+import { checkIfFileExists } from "../../src/utils";
 
 // Setup function that is not exported.
 const customViews = rewire("../../src/features/RunCode");
@@ -22,29 +21,27 @@ describe("RunCode feature", function () {
     before(utils.ensureEditorServicesIsConnected);
 
     it("Creates the launch config", function () {
-        const commandToRun: string = "Invoke-Build";
+        const commandToRun = "Invoke-Build";
         const args: string[] = ["Clean"];
 
         const expected: object = {
             request: "launch",
             type: "PowerShell",
-            name: "PowerShell Run Code",
-            script: commandToRun,
-            args,
+            name: "PowerShell: Run Code",
             internalConsoleOptions: "neverOpen",
             noDebug: false,
             createTemporaryIntegratedConsole: false,
-            cwd: vscode.workspace.rootPath,
+            script: commandToRun,
+            args,
         };
 
         const actual: object = createLaunchConfig(LaunchType.Debug, commandToRun, args);
-
         assert.deepStrictEqual(actual, expected);
     });
 
     it("Runs Pester tests from a file", async function () {
         const pesterTests = path.resolve(__dirname, "../../../examples/Tests/SampleModule.Tests.ps1");
-        assert(fs.existsSync(pesterTests));
+        assert(checkIfFileExists(pesterTests));
 
         // Open the PowerShell file with Pester tests and then wait a while for
         // the extension to finish connecting to the server.
