@@ -28,10 +28,9 @@ async function main() {
             version: "insiders"
         });
     } catch (err) {
-        // tslint:disable-next-line:no-console
         console.error(`Failed to run tests: ${err}`);
         process.exit(1);
     }
 }
 
-main();
+void main();
